test: clear the seven size and complexity outliers in test files
Follows the calibrated test-file thresholds by fixing what they flag, rather
than leaving the debt in place. No source file changes, so coverage is
unaffected: every package reports the same numbers as before.

Two cyclomatic-complexity outliers, both genuinely convoluted rather than
mismeasured:

  react-wizard  WizardFormIntegration validateFormData (27) — ten sequential
                `if (shouldValidate(x) && !values.x?.trim())` blocks replaced
                by a rule table walked in order, so `required` still takes
                precedence over `pattern` and `minLength`
  components    Menu panel-position assertions (21) — the ancestor walk and
                its inline-style fallbacks extracted into findFixedPanel and
                readInlineStyle, which also de-duplicates the identical walk
                in the sibling test

Five oversized suites split along the boundaries they already had. Shared
preambles moved into non-test helper modules; `jest.mock` factories stay in
each file, since babel-jest hoists them above that file's own imports:

  MapLibreAdapter.test.ts  1645 -> lifecycle + .translation + .mapData
  GeovisWorkspace.test.tsx 1403 -> layout + .slots + .warnings + .inspector
  react.test.tsx           1117 -> provider + .colorMode + .tokens
  GeoVisLegend.test.tsx    1058 -> rendering + .labelFormat
  emailAuth.test.ts        1033 -> config/password + .tokens + .flows

Test counts are preserved exactly across every split (73, 60, 66, 39 and 71
respectively), and each package's suite passes with unchanged coverage:
auth-core 261, components 260, fsl-theme 1295, geovis 789, geovis-workspace
96, react-wizard 49.

Repo-wide, max-lines drops from 9 to 4 and complexity from 31 to 29; the
remainder are pre-existing violations in `src`, left alone.

Note the auth-core PBKDF2 test timeout is unrelated to this and still open:
the 5 s limit is per test, so splitting the file does not affect it.
