+++ packages/fsl-theme/tests/unit/tests/engine/react/react.test.tsx
@@ -2,26 +2,20 @@
  * @jest-environment jsdom
  */
 
+/**
+ * ThemeProvider itself: mounting, prop reactivity, SSR style injection and the
+ * root target. Colour-mode behaviour lives in react.colorMode.test.tsx and
+ * token/style output in react.tokens.test.tsx.
+ */
+
 import { act, render, renderHook } from '@ttoss/test-utils/react';
 import type * as React from 'react';
 
 import { baseBundle } from '../../../../../src/baseBundle';
-import { createTheme } from '../../../../../src/createTheme';
 import {
-  getPreflightStyles,
-  PREFLIGHT_CSS,
-  toCssVars,
-} from '../../../../../src/css';
-import { useDatavizTokens } from '../../../../../src/dataviz/useDatavizTokens';
-import { withDataviz } from '../../../../../src/dataviz/withDataviz';
-import {
-  ThemeHead,
   ThemeProvider,
-  ThemeReset,
   ThemeScript,
-  ThemeStyles,
   useColorMode,
-  useResolvedTokens,
   useTokens,
 } from '../../../../../src/react';
 import { DATA_MODE_ATTR, DATA_THEME_ATTR } from '../../../../../src/runtime';
@@ -36,11 +30,6 @@ Object.defineProperty(window, 'matchMedia', {
 });
 
 const defaultBundle = baseBundle;
-const defaultWithDataviz = withDataviz(baseBundle);
-
-// ---------------------------------------------------------------------------
-// ThemeProvider
-// ---------------------------------------------------------------------------
 
 describe('ThemeProvider', () => {
   afterEach(clearDom);
@@ -125,10 +114,6 @@ describe('ThemeProvider', () => {
   });
 });
 
-// ---------------------------------------------------------------------------
-// ThemeScript
-// ---------------------------------------------------------------------------
-
 describe('ThemeScript', () => {
   test('renders a script tag', () => {
     const { container } = render(<ThemeScript />);
@@ -142,10 +127,6 @@ describe('ThemeScript', () => {
     expect(script?.getAttribute('nonce')).toBe('abc123');
   });
 });
-
-// ---------------------------------------------------------------------------
-// ThemeProvider prop reactivity
-// ---------------------------------------------------------------------------
 
 describe('ThemeProvider prop reactivity', () => {
   afterEach(clearDom);
@@ -262,11 +243,6 @@ describe('ThemeProvider prop reactivity', () => {
   });
 });
 
-// ---------------------------------------------------------------------------
-// SSR style injection — hoistable <style> carries a stable href (React 19
-// dedup key). Without href the same :root block duplicates per provider.
-// ---------------------------------------------------------------------------
-
 describe('SSR style injection (href dedup key)', () => {
   test('ThemeProvider server-renders a hoistable <style> keyed on a stable href + the CSS', async () => {
     const { renderToStaticMarkup } = await import('react-dom/server');
@@ -293,241 +269,6 @@ describe('SSR style injection (href dedup key)', () => {
   });
 });
 
-// ---------------------------------------------------------------------------
-// useTokens
-// ---------------------------------------------------------------------------
-
-describe('useTokens', () => {
-  afterEach(clearDom);
-
-  test('returns semantic tokens for the active theme', () => {
-    const { result } = renderHook(
-      () => {
-        return useTokens();
-      },
-      {
-        wrapper: ({ children }) => {
-          return (
-            <ThemeProvider defaultMode="light" theme={defaultBundle}>
-              {children}
-            </ThemeProvider>
-          );
-        },
-      }
-    );
-
-    expect(result.current.colors).toBeDefined();
-    expect(result.current.elevation).toBeDefined();
-    expect(result.current.text).toBeDefined();
-    expect(result.current.spacing).toBeDefined();
-  });
-
-  test('semantic tokens do not expose core', () => {
-    const { result } = renderHook(
-      () => {
-        return useTokens();
-      },
-      {
-        wrapper: ({ children }) => {
-          return (
-            <ThemeProvider defaultMode="light" theme={defaultBundle}>
-              {children}
-            </ThemeProvider>
-          );
-        },
-      }
-    );
-
-    // SemanticTokens type has no `core` property — verify at runtime too
-    expect((result.current as Record<string, unknown>)['core']).toBeUndefined();
-  });
-
-  test('returns dark-mode semantic tokens when resolvedMode is dark', () => {
-    const { result } = renderHook(
-      () => {
-        return useTokens();
-      },
-      {
-        wrapper: ({ children }) => {
-          return (
-            <ThemeProvider defaultMode="dark" theme={defaultBundle}>
-              {children}
-            </ThemeProvider>
-          );
-        },
-      }
-    );
-
-    // Dark mode remaps informational.primary.background to a dark neutral
-    const bg =
-      result.current.colors.informational?.primary?.background?.default;
-    expect(bg).not.toBe(
-      defaultBundle.base.semantic.colors.informational?.primary?.background
-        ?.default
-    );
-  });
-
-  test('throws when theme prop is missing', () => {
-    const consoleSpy = jest
-      .spyOn(console, 'error')
-      .mockImplementation(() => {});
-
-    expect(() => {
-      renderHook(
-        () => {
-          return useTokens();
-        },
-        {
-          wrapper: ({ children }) => {
-            return (
-              <ThemeProvider defaultMode="light">{children}</ThemeProvider>
-            );
-          },
-        }
-      );
-    }).toThrow('useTokens requires a <ThemeProvider theme={...}>');
-
-    consoleSpy.mockRestore();
-  });
-
-  test('throws when used outside ThemeProvider', () => {
-    const consoleSpy = jest
-      .spyOn(console, 'error')
-      .mockImplementation(() => {});
-
-    expect(() => {
-      renderHook(() => {
-        return useTokens();
-      });
-    }).toThrow('useTokens must be used within a <ThemeProvider>');
-
-    consoleSpy.mockRestore();
-  });
-
-  test('light→dark→light: returns to original token values', () => {
-    // Verifies resolveSemanticTokens + deepMerge idempotency through the
-    // full React context cycle.
-    const { result } = renderHook(
-      () => {
-        return { theme: useColorMode(), tokens: useTokens() };
-      },
-      {
-        wrapper: ({ children }) => {
-          return (
-            <ThemeProvider defaultMode="light" theme={defaultBundle}>
-              {children}
-            </ThemeProvider>
-          );
-        },
-      }
-    );
-
-    const lightBg =
-      result.current.tokens.colors.informational?.primary?.background?.default;
-
-    act(() => {
-      result.current.theme.setMode('dark');
-    });
-    const darkBg =
-      result.current.tokens.colors.informational?.primary?.background?.default;
-    expect(darkBg).not.toBe(lightBg);
-
-    act(() => {
-      result.current.theme.setMode('light');
-    });
-    const restoredBg =
-      result.current.tokens.colors.informational?.primary?.background?.default;
-    expect(restoredBg).toBe(lightBg);
-  });
-});
-
-// ---------------------------------------------------------------------------
-// useDatavizTokens (Phase 8.1)
-// ---------------------------------------------------------------------------
-
-describe('useDatavizTokens', () => {
-  afterEach(clearDom);
-
-  const datavizBundle = defaultWithDataviz;
-
-  test('returns dataviz tokens for theme with dataviz extension', () => {
-    const { result } = renderHook(
-      () => {
-        return useDatavizTokens();
-      },
-      {
-        wrapper: ({ children }) => {
-          return (
-            <ThemeProvider defaultMode="light" theme={datavizBundle}>
-              {children}
-            </ThemeProvider>
-          );
-        },
-      }
-    );
-
-    expect(result.current.color).toBeDefined();
-    expect(result.current.color.series).toBeDefined();
-    expect(result.current.color.series[1]).toBeDefined();
-    expect(result.current.encoding).toBeDefined();
-    expect(result.current.geo).toBeDefined();
-  });
-
-  test('throws descriptive error when theme has no dataviz extension', () => {
-    const consoleSpy = jest
-      .spyOn(console, 'error')
-      .mockImplementation(() => {});
-
-    expect(() => {
-      renderHook(
-        () => {
-          return useDatavizTokens();
-        },
-        {
-          wrapper: ({ children }) => {
-            return (
-              <ThemeProvider defaultMode="light" theme={defaultBundle}>
-                {children}
-              </ThemeProvider>
-            );
-          },
-        }
-      );
-    }).toThrow(
-      /useDatavizTokens: the active theme does not include the dataviz extension/
-    );
-
-    consoleSpy.mockRestore();
-  });
-
-  test('throws when theme prop is missing', () => {
-    const consoleSpy = jest
-      .spyOn(console, 'error')
-      .mockImplementation(() => {});
-
-    expect(() => {
-      renderHook(
-        () => {
-          return useDatavizTokens();
-        },
-        {
-          wrapper: ({ children }) => {
-            return (
-              <ThemeProvider defaultMode="light">{children}</ThemeProvider>
-            );
-          },
-        }
-      );
-    }).toThrow('useTokens requires a <ThemeProvider theme={...}>');
-
-    consoleSpy.mockRestore();
-  });
-});
-
-// ---------------------------------------------------------------------------
-// theme prop
-// ---------------------------------------------------------------------------
-
 describe('ThemeProvider theme prop', () => {
   afterEach(clearDom);
 
@@ -549,468 +290,6 @@ describe('ThemeProvider theme prop', () => {
     expect(result.current.elevation).toBeDefined();
   });
 });
-
-// ---------------------------------------------------------------------------
-// useColorMode
-// ---------------------------------------------------------------------------
-
-describe('useColorMode', () => {
-  afterEach(clearDom);
-
-  test('returns mode state', () => {
-    const { result } = renderHook(
-      () => {
-        return useColorMode();
-      },
-      {
-        wrapper: ({ children }) => {
-          return <ThemeProvider defaultMode="light">{children}</ThemeProvider>;
-        },
-      }
-    );
-
-    expect(result.current.mode).toBe('light');
-    expect(result.current.resolvedMode).toBe('light');
-    expect(typeof result.current.setMode).toBe('function');
-  });
-
-  test('setMode updates mode state', () => {
-    const { result } = renderHook(
-      () => {
-        return useColorMode();
-      },
-      {
-        wrapper: ({ children }) => {
-          return <ThemeProvider defaultMode="light">{children}</ThemeProvider>;
-        },
-      }
-    );
-
-    act(() => {
-      result.current.setMode('dark');
-    });
-
-    expect(result.current.mode).toBe('dark');
-    expect(result.current.resolvedMode).toBe('dark');
-  });
-
-  test('throws when used outside ThemeProvider', () => {
-    const consoleSpy = jest
-      .spyOn(console, 'error')
-      .mockImplementation(() => {});
-
-    expect(() => {
-      renderHook(() => {
-        return useColorMode();
-      });
-    }).toThrow('useColorMode must be used within a <ThemeProvider>');
-
-    consoleSpy.mockRestore();
-  });
-});
-
-// ---------------------------------------------------------------------------
-// useResolvedTokens
-// ---------------------------------------------------------------------------
-
-describe('useResolvedTokens', () => {
-  afterEach(clearDom);
-
-  test('returns flat map of semantic.* keys with resolved values', () => {
-    const { result } = renderHook(
-      () => {
-        return useResolvedTokens();
-      },
-      {
-        wrapper: ({ children }) => {
-          return (
-            <ThemeProvider defaultMode="light" theme={defaultBundle}>
-              {children}
-            </ThemeProvider>
-          );
-        },
-      }
-    );
-
-    const keys = Object.keys(result.current);
-    expect(keys.length).toBeGreaterThan(0);
-    for (const key of keys) {
-      expect(key.startsWith('semantic.')).toBe(true);
-    }
-  });
-
-  test('values are raw (not unresolved token refs)', () => {
-    const { result } = renderHook(
-      () => {
-        return useResolvedTokens();
-      },
-      {
-        wrapper: ({ children }) => {
-          return (
-            <ThemeProvider defaultMode="light" theme={defaultBundle}>
-              {children}
-            </ThemeProvider>
-          );
-        },
-      }
-    );
-
-    for (const value of Object.values(result.current)) {
-      if (typeof value === 'string') {
-        expect(value).not.toMatch(/^\{.+\}$/);
-      }
-    }
-  });
-
-  test('dark mode returns different resolved values for mode-sensitive tokens', () => {
-    const { result: lightResult } = renderHook(
-      () => {
-        return useResolvedTokens();
-      },
-      {
-        wrapper: ({ children }) => {
-          return (
-            <ThemeProvider defaultMode="light" theme={defaultBundle}>
-              {children}
-            </ThemeProvider>
-          );
-        },
-      }
-    );
-
-    const { result: darkResult } = renderHook(
-      () => {
-        return useResolvedTokens();
-      },
-      {
-        wrapper: ({ children }) => {
-          return (
-            <ThemeProvider defaultMode="dark" theme={defaultBundle}>
-              {children}
-            </ThemeProvider>
-          );
-        },
-      }
-    );
-
-    const lightValues = lightResult.current;
-    const darkValues = darkResult.current;
-    const allKeys = Object.keys(lightValues);
-    const hasDiff = allKeys.some((k) => {
-      return lightValues[k] !== darkValues[k];
-    });
-    expect(hasDiff).toBe(true);
-  });
-
-  test('throws when theme prop is missing', () => {
-    const consoleSpy = jest
-      .spyOn(console, 'error')
-      .mockImplementation(() => {});
-
-    expect(() => {
-      renderHook(
-        () => {
-          return useResolvedTokens();
-        },
-        {
-          wrapper: ({ children }) => {
-            return (
-              <ThemeProvider defaultMode="light">{children}</ThemeProvider>
-            );
-          },
-        }
-      );
-    }).toThrow('useResolvedTokens requires a <ThemeProvider theme={...}>');
-
-    consoleSpy.mockRestore();
-  });
-
-  test('throws when used outside ThemeProvider', () => {
-    const consoleSpy = jest
-      .spyOn(console, 'error')
-      .mockImplementation(() => {});
-
-    expect(() => {
-      renderHook(() => {
-        return useResolvedTokens();
-      });
-    }).toThrow('useResolvedTokens must be used within a <ThemeProvider>');
-
-    consoleSpy.mockRestore();
-  });
-
-  test('applies coarse-pointer hit overrides when any-pointer: coarse matches', () => {
-    // Simulate coarse pointer (touch device)
-    jest
-      .mocked(window.matchMedia)
-      .mockImplementation(matchMediaMockImpl({ coarsePointer: true }));
-
-    const { result } = renderHook(
-      () => {
-        return useResolvedTokens();
-      },
-      {
-        wrapper: ({ children }) => {
-          return (
-            <ThemeProvider defaultMode="light" theme={defaultBundle}>
-              {children}
-            </ThemeProvider>
-          );
-        },
-      }
-    );
-
-    // Coarse hit value should be the raw core.sizing.hit.coarse value
-    expect(result.current['semantic.sizing.hit']).toBe(
-      defaultBundle.base.core.sizing.hit.coarse
-    );
-
-    // Restore default mock
-    jest.mocked(window.matchMedia).mockImplementation(matchMediaMockImpl());
-  });
-
-  test('uses fine-pointer hit values when any-pointer: coarse does not match', () => {
-    // Default mock: no coarse pointer
-    jest.mocked(window.matchMedia).mockImplementation(matchMediaMockImpl());
-
-    const { result } = renderHook(
-      () => {
-        return useResolvedTokens();
-      },
-      {
-        wrapper: ({ children }) => {
-          return (
-            <ThemeProvider defaultMode="light" theme={defaultBundle}>
-              {children}
-            </ThemeProvider>
-          );
-        },
-      }
-    );
-
-    // Fine hit value — a clamp() expression resolved from core.sizing.hit.fine
-    expect(result.current['semantic.sizing.hit']).toBe(
-      defaultBundle.base.core.sizing.hit.fine
-    );
-  });
-});
-
-// ---------------------------------------------------------------------------
-// onModeChange callback
-// ---------------------------------------------------------------------------
-
-describe('ThemeProvider onModeChange', () => {
-  afterEach(clearDom);
-
-  test('does NOT fire on initial mount', () => {
-    const onModeChange = jest.fn();
-
-    renderHook(
-      () => {
-        return useColorMode();
-      },
-      {
-        wrapper: ({ children }) => {
-          return (
-            <ThemeProvider defaultMode="light" onModeChange={onModeChange}>
-              {children}
-            </ThemeProvider>
-          );
-        },
-      }
-    );
-
-    expect(onModeChange).not.toHaveBeenCalled();
-  });
-
-  test('fires after setMode is called', () => {
-    const onModeChange = jest.fn();
-
-    const { result } = renderHook(
-      () => {
-        return useColorMode();
-      },
-      {
-        wrapper: ({ children }) => {
-          return (
-            <ThemeProvider defaultMode="light" onModeChange={onModeChange}>
-              {children}
-            </ThemeProvider>
-          );
-        },
-      }
-    );
-
-    act(() => {
-      result.current.setMode('dark');
-    });
-
-    expect(onModeChange).toHaveBeenCalledTimes(1);
-    expect(onModeChange).toHaveBeenCalledWith('dark', 'dark');
-  });
-
-  test('receives both mode and resolvedMode', () => {
-    const onModeChange = jest.fn();
-
-    const { result } = renderHook(
-      () => {
-        return useColorMode();
-      },
-      {
-        wrapper: ({ children }) => {
-          return (
-            <ThemeProvider defaultMode="light" onModeChange={onModeChange}>
-              {children}
-            </ThemeProvider>
-          );
-        },
-      }
-    );
-
-    act(() => {
-      result.current.setMode('dark');
-    });
-
-    const [mode, resolvedMode] = onModeChange.mock.calls[0];
-    expect(mode).toBe('dark');
-    expect(resolvedMode).toBe('dark');
-  });
-
-  test('fires again on subsequent mode changes', () => {
-    const onModeChange = jest.fn();
-
-    const { result } = renderHook(
-      () => {
-        return useColorMode();
-      },
-      {
-        wrapper: ({ children }) => {
-          return (
-            <ThemeProvider defaultMode="light" onModeChange={onModeChange}>
-              {children}
-            </ThemeProvider>
-          );
-        },
-      }
-    );
-
-    act(() => {
-      result.current.setMode('dark');
-    });
-    act(() => {
-      result.current.setMode('light');
-    });
-
-    expect(onModeChange).toHaveBeenCalledTimes(2);
-    expect(onModeChange.mock.calls[1]).toEqual(['light', 'light']);
-  });
-
-  test('onModeChange does NOT fire when root prop changes from undefined to element', () => {
-    const onModeChange = jest.fn();
-    const container = document.createElement('div');
-    document.body.appendChild(container);
-
-    const Wrapper = ({ root }: { root: HTMLElement | undefined }) => {
-      return (
-        <ThemeProvider
-          defaultMode="light"
-          theme={defaultBundle}
-          root={root}
-          onModeChange={onModeChange}
-        >
-          <div />
-        </ThemeProvider>
-      );
-    };
-
-    const { rerender, unmount } = render(<Wrapper root={undefined} />);
-
-    // Simulate ref becoming available (Storybook pattern: undefined → element)
-    rerender(<Wrapper root={container} />);
-
-    // onModeChange must NOT fire — no user action, no mode change
-    expect(onModeChange).not.toHaveBeenCalled();
-
-    unmount();
-    container.remove();
-  });
-
-  test('onModeChange fires after mode change following root transition', () => {
-    const onModeChange = jest.fn();
-    const container = document.createElement('div');
-    document.body.appendChild(container);
-
-    const { result, rerender, unmount } = renderHook(
-      () => {
-        return useColorMode();
-      },
-      {
-        wrapper: ({ children }) => {
-          return (
-            <ThemeProvider
-              defaultMode="light"
-              theme={defaultBundle}
-              root={container}
-              onModeChange={onModeChange}
-            >
-              <div />
-              {children}
-            </ThemeProvider>
-          );
-        },
-      }
-    );
-
-    // Trigger a re-render (simulating root transition)
-    rerender();
-
-    // Now trigger a real mode change
-    act(() => {
-      result.current.setMode('dark');
-    });
-
-    expect(onModeChange).toHaveBeenCalledTimes(1);
-    expect(onModeChange).toHaveBeenCalledWith('dark', 'dark');
-
-    unmount();
-    container.remove();
-  });
-
-  test('onModeChange does NOT fire when root changes multiple times without mode change', () => {
-    const onModeChange = jest.fn();
-    const container1 = document.createElement('div');
-    const container2 = document.createElement('div');
-    document.body.appendChild(container1);
-    document.body.appendChild(container2);
-
-    const Wrapper = ({ root }: { root: HTMLElement | undefined }) => {
-      return (
-        <ThemeProvider
-          defaultMode="dark"
-          theme={defaultBundle}
-          root={root}
-          onModeChange={onModeChange}
-        >
-          <div />
-        </ThemeProvider>
-      );
-    };
-
-    const { rerender, unmount } = render(<Wrapper root={undefined} />);
-    rerender(<Wrapper root={container1} />);
-    rerender(<Wrapper root={container2} />);
-
-    expect(onModeChange).not.toHaveBeenCalled();
-
-    unmount();
-    container1.remove();
-    container2.remove();
-  });
-});
-
-// ---------------------------------------------------------------------------
-// root prop
-// ---------------------------------------------------------------------------
 
 describe('ThemeProvider root prop', () => {
   afterEach(clearDom);
@@ -1064,259 +343,6 @@ describe('ThemeProvider root prop', () => {
   });
 });
 
-// ---------------------------------------------------------------------------
-// ThemeStyles
-// ---------------------------------------------------------------------------
-
-describe('ThemeStyles', () => {
-  test('renders a style tag', () => {
-    const { container } = render(
-      <ThemeStyles theme={defaultBundle} themeId="default" />
-    );
-    const style = container.querySelector('style');
-    expect(style).not.toBeNull();
-  });
-
-  test('style scoped to themeId selector', () => {
-    const { container } = render(
-      <ThemeStyles theme={defaultBundle} themeId="default" />
-    );
-    const style = container.querySelector('style');
-    expect(style?.innerHTML).toContain('[data-tt-theme="default"]');
-  });
-
-  test('style targets :root when themeId is omitted', () => {
-    const { container } = render(<ThemeStyles theme={defaultBundle} />);
-    const style = container.querySelector('style');
-    expect(style?.innerHTML).toContain(':root {');
-    expect(style?.innerHTML).not.toContain('[data-tt-theme');
-  });
-
-  test('alternate mode uses :root[data-tt-mode] when themeId is omitted', () => {
-    const { container } = render(<ThemeStyles theme={defaultBundle} />);
-    const style = container.querySelector('style');
-    expect(style?.innerHTML).toContain(':root[data-tt-mode="dark"]');
-  });
-
-  test('passes nonce to style tag', () => {
-    const { container } = render(
-      <ThemeStyles theme={defaultBundle} themeId="default" nonce="abc123" />
-    );
-    const style = container.querySelector('style');
-    expect(style?.getAttribute('nonce')).toBe('abc123');
-  });
-});
-
-// ---------------------------------------------------------------------------
-// ThemeHead
-// ---------------------------------------------------------------------------
-
-describe('ThemeHead', () => {
-  test('renders a script tag and a style tag', () => {
-    const { container } = render(<ThemeHead theme={defaultBundle} />);
-    expect(container.querySelector('script')).not.toBeNull();
-    expect(container.querySelector('style')).not.toBeNull();
-  });
-
-  test('style targets :root when themeId is omitted', () => {
-    const { container } = render(<ThemeHead theme={defaultBundle} />);
-    const style = container.querySelector('style');
-    expect(style?.innerHTML).toContain(':root {');
-    expect(style?.innerHTML).not.toContain('[data-tt-theme');
-  });
-
-  test('style scoped to themeId when provided', () => {
-    const { container } = render(
-      <ThemeHead theme={defaultBundle} themeId="default" />
-    );
-    const style = container.querySelector('style');
-    expect(style?.innerHTML).toContain('[data-tt-theme="default"]');
-  });
-
-  test('passes nonce to both script and style', () => {
-    const { container } = render(
-      <ThemeHead theme={defaultBundle} nonce="xyz" />
-    );
-    expect(container.querySelector('script')?.getAttribute('nonce')).toBe(
-      'xyz'
-    );
-    expect(container.querySelector('style')?.getAttribute('nonce')).toBe('xyz');
-  });
-
-  test('passes defaultMode to ThemeScript', () => {
-    const { container } = render(
-      <ThemeHead theme={defaultBundle} defaultMode="dark" />
-    );
-    const script = container.querySelector('script');
-    expect(script?.innerHTML).toContain('"dark"');
-  });
-});
-
-// ---------------------------------------------------------------------------
-// DEV-only warnings — root/themeId pairing and hoisted-style dedup mismatch
-// ---------------------------------------------------------------------------
-
-describe('ThemeProvider DEV warnings', () => {
-  let warnSpy: jest.SpyInstance;
-
-  beforeEach(() => {
-    warnSpy = jest.spyOn(console, 'warn').mockImplementation(() => {});
-  });
-
-  afterEach(() => {
-    warnSpy.mockRestore();
-    clearDom();
-  });
-
-  test('warns when root is passed without themeId', () => {
-    const container = document.createElement('div');
-    document.body.appendChild(container);
-
-    render(
-      <ThemeProvider theme={defaultBundle} root={container}>
-        <div>child</div>
-      </ThemeProvider>
-    );
-
-    expect(warnSpy).toHaveBeenCalledWith(
-      expect.stringContaining('`root` was passed without `themeId`')
-    );
-    container.remove();
-  });
-
-  test('does not warn when root is paired with themeId', () => {
-    const container = document.createElement('div');
-    document.body.appendChild(container);
-
-    render(
-      <ThemeProvider theme={defaultBundle} themeId="scoped" root={container}>
-        <div>child</div>
-      </ThemeProvider>
-    );
-
-    expect(warnSpy).not.toHaveBeenCalledWith(
-      expect.stringContaining('`root` was passed without `themeId`')
-    );
-    container.remove();
-  });
-
-  test('warns when two providers with different themes share the same style href', () => {
-    const themeA = createTheme({
-      overrides: { core: { colors: { brand: { 500: '#AA0000' } } } },
-    });
-    const themeB = createTheme({
-      overrides: { core: { colors: { brand: { 500: '#00BB00' } } } },
-    });
-
-    render(
-      <>
-        <ThemeProvider theme={themeA}>
-          <div>a</div>
-        </ThemeProvider>
-        <ThemeProvider theme={themeB}>
-          <div>b</div>
-        </ThemeProvider>
-      </>
-    );
-
-    expect(warnSpy).toHaveBeenCalledWith(
-      expect.stringContaining('share the same style key')
-    );
-  });
-
-  test('does not warn for two providers with distinct themeIds', () => {
-    const themeA = createTheme({
-      overrides: { core: { colors: { brand: { 500: '#AA0000' } } } },
-    });
-    const themeB = createTheme({
-      overrides: { core: { colors: { brand: { 500: '#00BB00' } } } },
-    });
-
-    render(
-      <>
-        <ThemeProvider theme={themeA} themeId="brand-a">
-          <div>a</div>
-        </ThemeProvider>
-        <ThemeProvider theme={themeB} themeId="brand-b">
-          <div>b</div>
-        </ThemeProvider>
-      </>
-    );
-
-    expect(warnSpy).not.toHaveBeenCalledWith(
-      expect.stringContaining('share the same style key')
-    );
-  });
-
-  test('does not warn when the same theme is re-rendered (dedup is intended)', () => {
-    const { rerender } = render(
-      <ThemeProvider theme={defaultBundle}>
-        <div>a</div>
-      </ThemeProvider>
-    );
-    rerender(
-      <ThemeProvider theme={defaultBundle}>
-        <div>b</div>
-      </ThemeProvider>
-    );
-
-    expect(warnSpy).not.toHaveBeenCalledWith(
-      expect.stringContaining('share the same style key')
-    );
-  });
-});
-
-// ---------------------------------------------------------------------------
-// systemModeFallback derivation from defaultMode
-// ---------------------------------------------------------------------------
-
-describe('OS-preference fallback follows defaultMode', () => {
-  // renderToStaticMarkup sidesteps React 19's per-document hoisted-style
-  // cache, which would swallow repeat injections of the same href in jsdom.
-  const staticMarkup = (node: React.ReactElement): string => {
-    // eslint-disable-next-line @typescript-eslint/no-require-imports
-    const { renderToStaticMarkup } = require('react-dom/server');
-    return renderToStaticMarkup(node);
-  };
-
-  test('ThemeProvider: system default emits the fallback; fixed light does not', () => {
-    const system = staticMarkup(
-      <ThemeProvider theme={defaultBundle}>
-        <div>x</div>
-      </ThemeProvider>
-    );
-    const light = staticMarkup(
-      <ThemeProvider theme={defaultBundle} defaultMode="light">
-        <div>x</div>
-      </ThemeProvider>
-    );
-
-    expect(system).toContain('@media (prefers-color-scheme: dark)');
-    expect(light).not.toContain('@media (prefers-color-scheme:');
-  });
-
-  test('ThemeHead derives the gate from its defaultMode', () => {
-    const fixedDark = staticMarkup(
-      <ThemeHead theme={defaultBundle} defaultMode="dark" />
-    );
-    const system = staticMarkup(<ThemeHead theme={defaultBundle} />);
-
-    expect(fixedDark).not.toContain('@media (prefers-color-scheme:');
-    expect(system).toContain('@media (prefers-color-scheme: dark)');
-  });
-
-  test('ThemeStyles exposes an explicit systemModeFallback prop', () => {
-    const suppressed = staticMarkup(
-      <ThemeStyles theme={defaultBundle} systemModeFallback={false} />
-    );
-    expect(suppressed).not.toContain('@media (prefers-color-scheme:');
-  });
-});
-
-// ---------------------------------------------------------------------------
-// root as RefObject — no transient attach to <html>
-// ---------------------------------------------------------------------------
-
 describe('ThemeProvider root as RefObject', () => {
   let warnSpy: jest.SpyInstance;
 
@@ -1361,62 +387,6 @@ describe('ThemeProvider root as RefObject', () => {
     // No spurious multi-runtime warning — the ref form never attaches to <html>.
     expect(warnSpy).not.toHaveBeenCalledWith(
       expect.stringContaining('Multiple theme runtimes')
-    );
-  });
-});
-
-describe('ThemeReset / preflight', () => {
-  test('getPreflightStyles returns the preflight CSS', () => {
-    expect(getPreflightStyles()).toBe(PREFLIGHT_CSS);
-  });
-
-  test('preflight resets box-sizing and binds the body to tokens', () => {
-    expect(PREFLIGHT_CSS).toContain('box-sizing: border-box;');
-    expect(PREFLIGHT_CSS).toContain(
-      'var(--tt-colors-informational-primary-background-default)'
-    );
-    expect(PREFLIGHT_CSS).toContain('prefers-reduced-motion: reduce');
-    // Layout-agnostic: the base declares no layout (that is fsl-ui / the app).
-    expect(PREFLIGHT_CSS).not.toContain('display:');
-    expect(PREFLIGHT_CSS).not.toContain('grid');
-  });
-
-  test('every var() the preflight reads is emitted by the theme CSS', () => {
-    // Guard against emitter/preflight naming drift: the preflight is a static
-    // string, so a rename in the emitter (toCssVarName) would otherwise make
-    // its var() reads silently fall back (T-20). Every custom property the
-    // preflight reads must exist in the vars record the emitter produces.
-    const emittedVars = toCssVars(defaultBundle).base.cssVars;
-    const readVars = [...PREFLIGHT_CSS.matchAll(/var\((--tt-[\w-]+)/g)].map(
-      (match) => {
-        return match[1];
-      }
-    );
-
-    // Sanity: the body typography reads are among them.
-    expect(readVars).toEqual(
-      expect.arrayContaining([
-        '--tt-text-body-md-fontFamily',
-        '--tt-text-body-md-fontSize',
-        '--tt-text-body-md-lineHeight',
-      ])
-    );
-
-    for (const varName of readVars) {
-      expect(emittedVars).toHaveProperty([varName]);
-    }
-  });
-
-  test('ThemeReset injects the preflight into a <style> tag', () => {
-    const { container } = render(<ThemeReset />);
-    const style = container.querySelector('style');
-    expect(style?.textContent).toBe(PREFLIGHT_CSS);
-  });
-
-  test('ThemeReset forwards a CSP nonce', () => {
-    const { container } = render(<ThemeReset nonce="abc123" />);
-    expect(container.querySelector('style')?.getAttribute('nonce')).toBe(
-      'abc123'
     );
   });
 });
