+++ packages/geovis-workspace/tests/unit/tests/GeovisWorkspace.test.tsx
@@ -1,212 +1,25 @@
+/**
+ * Workspace layout, the left sidebar and variable selection. Right-sidebar
+ * slots, the warnings panel and the inspector/detail API each have their own
+ * suite alongside this one.
+ */
+
 import { useGeoVis } from '@ttoss/geovis';
-import { I18nProvider } from '@ttoss/react-i18n';
 import { act, fireEvent, render, screen } from '@ttoss/test-utils/react';
-import type * as React from 'react';
-import {
-  GeovisWorkspace,
-  type GeovisWorkspaceConfig,
-  GeovisWorkspaceProvider,
-  getInitialSelection,
-  useGeovisWorkspace,
-} from 'src';
+import { GeovisWorkspace, useGeovisWorkspace } from 'src';
 import { LEFT_SIDEBAR_CONTROL_CLEARANCE } from 'src/controlOffset';
 
-interface MockClick {
-  layerId: string;
-  featureId: string | number;
-  value: number | string | null;
-}
-
-interface MockSpec {
-  legends?: { id: string }[];
-  mockResult?: unknown;
-  mockClick?: MockClick | null;
-}
+import {
+  config,
+  openLeftSidebar,
+  Provider,
+  visualizationSpec,
+} from './geovisWorkspaceTestUtils';
 
 jest.mock('@ttoss/geovis', () => {
-  const ReactModule = jest.requireActual('react');
-  const MockGeoVisContext = ReactModule.createContext<{
-    spec: MockSpec;
-    result: unknown;
-    click: MockClick | null;
-    dismiss: () => void;
-  } | null>(null);
-
-  return {
-    GeoVisProvider: ({
-      spec,
-      children,
-    }: React.PropsWithChildren<{ spec: MockSpec }>) => {
-      const result = spec.mockResult ?? {
-        status: 'resolved',
-        spec,
-        warnings: [],
-      };
-
-      const [click, setClick] = ReactModule.useState<MockClick | null>(() => {
-        return spec.mockClick ?? null;
-      });
-
-      ReactModule.useEffect(() => {
-        setClick(spec.mockClick ?? null);
-      }, [spec.mockClick]);
-
-      const dismiss = ReactModule.useCallback(() => {
-        setClick(null);
-      }, []);
-
-      return (
-        <MockGeoVisContext.Provider value={{ spec, result, click, dismiss }}>
-          <div data-testid="geovis-provider">{children}</div>
-        </MockGeoVisContext.Provider>
-      );
-    },
-    GeoVisCanvas: () => {
-      return <div data-testid="geovis-canvas" />;
-    },
-    useGeoVis: () => {
-      const context = ReactModule.useContext(MockGeoVisContext);
-      if (!context) throw new Error('useGeoVis used outside GeoVisProvider');
-      return context;
-    },
-    useGeoVisClick: () => {
-      const context = ReactModule.useContext(MockGeoVisContext);
-      if (!context) {
-        throw new Error('useGeoVisClick used outside GeoVisProvider');
-      }
-      return context.click;
-    },
-    useDismissGeoVisClick: () => {
-      const context = ReactModule.useContext(MockGeoVisContext);
-      if (!context) {
-        throw new Error('useDismissGeoVisClick used outside GeoVisProvider');
-      }
-      return context.dismiss;
-    },
-    GeoVisLegend: ({ legendId }: { legendId: string }) => {
-      return <div data-testid={`legend-${legendId}`}>{legendId}</div>;
-    },
-  };
+  // eslint-disable-next-line @typescript-eslint/no-require-imports -- the factory is hoisted above imports
+  return require('./geovisWorkspaceTestUtils').createGeoVisMock();
 });
-
-const Provider = ({ children }: React.PropsWithChildren) => {
-  return <I18nProvider>{children}</I18nProvider>;
-};
-
-const config: GeovisWorkspaceConfig = {
-  leftSidebar: {
-    sections: [
-      {
-        id: 'population',
-        header: { title: 'População' },
-        body: {
-          kind: 'variations',
-          menuId: 'population',
-          groups: [
-            {
-              id: 'population',
-              label: 'População',
-              variations: [
-                { value: '5year-65plus', label: 'Faixa (% da pop 65+)' },
-                { value: '0-14', label: '0 a 14 anos' },
-              ],
-            },
-          ],
-        },
-      },
-      {
-        id: 'economy',
-        header: { title: 'Economia' },
-        body: {
-          kind: 'variations',
-          menuId: 'economy',
-          groups: [
-            {
-              id: 'economy',
-              label: 'Economia',
-              variations: [
-                { value: 'gdp', label: 'PIB' },
-                { value: 'income', label: 'Renda média' },
-              ],
-            },
-          ],
-        },
-      },
-    ],
-  },
-};
-
-const visualizationSpec = {
-  id: 'test-spec',
-  engine: 'maplibre' as const,
-  sources: [],
-  layers: [],
-};
-
-const visualizationSpecWithLegends = {
-  ...visualizationSpec,
-  legends: [{ id: 'classes' }],
-};
-
-const resolvedWithWarnings = {
-  status: 'resolved' as const,
-  spec: visualizationSpec,
-  warnings: [
-    {
-      code: 'policy-violation' as const,
-      subject: { path: 'metadata.metricField', id: 'policy-invalid' },
-      message: 'Spec violates policy.',
-      repair: [
-        {
-          kind: 'set-value' as const,
-          path: 'metadata.metricField',
-          value: 'safe-field',
-          label: "Use 'safe-field'",
-        },
-      ],
-    },
-  ],
-};
-
-const failingResult = {
-  status: 'mismatch' as const,
-  issues: [
-    {
-      code: 'unknown-map-data-id' as const,
-      subject: { path: 'layers[0].mapDataId', id: 'missing-id' },
-      message: 'Unknown map data id.',
-      repair: [
-        {
-          kind: 'allowed-values' as const,
-          path: 'layers[0].mapDataId',
-          values: ['choropleth', 'dots'],
-        },
-      ],
-    },
-  ],
-};
-
-const visualizationSpecWithWarnings = {
-  ...visualizationSpec,
-  mockResult: resolvedWithWarnings,
-};
-
-const failingVisualizationSpec = {
-  ...visualizationSpec,
-  mockResult: failingResult,
-};
-
-const openLeftSidebar = async () => {
-  await act(async () => {
-    fireEvent.click(screen.getByRole('button', { name: 'Open menu' }));
-  });
-};
-
-const openRightSidebar = async () => {
-  await act(async () => {
-    fireEvent.click(screen.getByRole('button', { name: 'Open details' }));
-  });
-};
 
 test('renders the GeoVis map canvas inside the main content area', () => {
   render(
@@ -487,1190 +300,4 @@ test('controlled variables prop drives the active item', async () => {
     'aria-pressed',
     'false'
   );
-});
-
-test('right sidebar renders only when a slot has content', () => {
-  const { rerender } = render(
-    <GeovisWorkspace config={config} visualizationSpec={visualizationSpec} />,
-    { wrapper: Provider }
-  );
-
-  expect(
-    screen.queryByRole('button', { name: 'Open details' })
-  ).not.toBeInTheDocument();
-
-  rerender(
-    <GeovisWorkspace
-      config={{ ...config, legend: { description: 'Descrição' } }}
-      visualizationSpec={visualizationSpec}
-    />
-  );
-
-  expect(
-    screen.getByRole('button', { name: 'Open details' })
-  ).toBeInTheDocument();
-});
-
-test('left sidebar controls are absent when leftSidebar has no sections', () => {
-  render(
-    <GeovisWorkspace
-      config={{ legend: { description: 'Descrição' } }}
-      visualizationSpec={visualizationSpec}
-    />,
-    { wrapper: Provider }
-  );
-
-  expect(
-    screen.queryByRole('button', { name: 'Open menu' })
-  ).not.toBeInTheDocument();
-});
-
-test('right sidebar shows a custom title', async () => {
-  render(
-    <GeovisWorkspace
-      config={{
-        ...config,
-        legend: { description: 'Descrição' },
-        rightSidebar: { title: 'Camadas' },
-      }}
-      visualizationSpec={visualizationSpec}
-    />,
-    { wrapper: Provider }
-  );
-
-  await openRightSidebar();
-
-  expect(screen.getByText('Camadas')).toBeInTheDocument();
-});
-
-test('right sidebar renders the legend panel from the config', async () => {
-  const configWithLegend: GeovisWorkspaceConfig = {
-    ...config,
-    rightSidebar: { title: 'População 65+' },
-    legend: {
-      description: 'Proporção da população total com 65 anos ou mais.',
-      sources: {
-        title: 'Fonte dos dados:',
-        items: [
-          { label: 'SEADE (2025)', href: 'https://example.com/seade' },
-          { label: 'Geometria: Distritos Municipais.' },
-        ],
-      },
-    },
-  };
-
-  render(
-    <GeovisWorkspace
-      config={configWithLegend}
-      visualizationSpec={visualizationSpecWithLegends}
-    />,
-    { wrapper: Provider }
-  );
-
-  await openRightSidebar();
-
-  expect(
-    screen.getByText('Proporção da população total com 65 anos ou mais.')
-  ).toBeInTheDocument();
-  // Spec legends are map overlays now — the legend slot no longer auto-dumps
-  // them into the right sidebar; it shows only the consumer's config.legend.
-  expect(screen.queryByTestId('legend-classes')).not.toBeInTheDocument();
-  expect(screen.getByText('Fonte dos dados:')).toBeInTheDocument();
-  expect(
-    screen.getByText('Geometria: Distritos Municipais.')
-  ).toBeInTheDocument();
-
-  const link = screen.getByRole('link', { name: 'SEADE (2025)' });
-  expect(link).toHaveAttribute('href', 'https://example.com/seade');
-  expect(link).toHaveAttribute('target', '_blank');
-});
-
-test('right sidebar shows when the legend slot only has sources configured', async () => {
-  render(
-    <GeovisWorkspace
-      config={{
-        ...config,
-        legend: { sources: { items: [{ label: 'SEADE (2025)' }] } },
-      }}
-      visualizationSpec={visualizationSpec}
-    />,
-    { wrapper: Provider }
-  );
-
-  await openRightSidebar();
-
-  expect(screen.getByText('SEADE (2025)')).toBeInTheDocument();
-});
-
-test('right sidebar is absent when the legend slot has no content and nothing else configures it', () => {
-  render(
-    <GeovisWorkspace
-      config={{ ...config, legend: {} }}
-      visualizationSpec={visualizationSpec}
-    />,
-    { wrapper: Provider }
-  );
-
-  expect(
-    screen.queryByRole('button', { name: 'Open details' })
-  ).not.toBeInTheDocument();
-});
-
-test('hiding the legend slot suppresses it even when the spec has legends', () => {
-  render(
-    <GeovisWorkspace
-      config={{ ...config, slots: { legend: { hidden: true } } }}
-      visualizationSpec={visualizationSpecWithLegends}
-    />,
-    { wrapper: Provider }
-  );
-
-  expect(
-    screen.queryByRole('button', { name: 'Open details' })
-  ).not.toBeInTheDocument();
-});
-
-test('a controls slot override replaces the default sidebar and keeps the sidebar visible', async () => {
-  const CustomControls = () => {
-    return <div data-testid="custom-controls">custom</div>;
-  };
-
-  render(
-    <GeovisWorkspace
-      config={{ slots: { controls: { component: CustomControls } } }}
-      visualizationSpec={visualizationSpec}
-    />,
-    { wrapper: Provider }
-  );
-
-  await openLeftSidebar();
-
-  expect(screen.getByTestId('custom-controls')).toBeInTheDocument();
-  expect(screen.queryByText('População')).not.toBeInTheDocument();
-});
-
-test('hiding the map slot renders no canvas', () => {
-  render(
-    <GeovisWorkspace
-      config={{ ...config, slots: { map: { hidden: true } } }}
-      visualizationSpec={visualizationSpec}
-    />,
-    { wrapper: Provider }
-  );
-
-  expect(screen.queryByTestId('geovis-canvas')).not.toBeInTheDocument();
-});
-
-test('a map slot override replaces the default canvas', () => {
-  const CustomMap = () => {
-    return <div data-testid="custom-map">custom map</div>;
-  };
-
-  render(
-    <GeovisWorkspace
-      config={{ ...config, slots: { map: { component: CustomMap } } }}
-      visualizationSpec={visualizationSpec}
-    />,
-    { wrapper: Provider }
-  );
-
-  expect(screen.getByTestId('custom-map')).toBeInTheDocument();
-  expect(screen.queryByTestId('geovis-canvas')).not.toBeInTheDocument();
-});
-
-test('right sidebar shows via a slot override even when legend has no config', async () => {
-  const CustomMetadata = () => {
-    return <div data-testid="custom-metadata">meta</div>;
-  };
-
-  render(
-    <GeovisWorkspace
-      config={{ ...config, slots: { metadata: { component: CustomMetadata } } }}
-      visualizationSpec={visualizationSpec}
-    />,
-    { wrapper: Provider }
-  );
-
-  await openRightSidebar();
-
-  expect(screen.getByTestId('custom-metadata')).toBeInTheDocument();
-});
-
-test('hidden takes precedence over an override for a right-sidebar slot', async () => {
-  const CustomMetadata = () => {
-    return <div data-testid="custom-metadata">meta</div>;
-  };
-
-  render(
-    <GeovisWorkspace
-      config={{
-        ...config,
-        legend: { description: 'Descrição' },
-        slots: { metadata: { component: CustomMetadata, hidden: true } },
-      }}
-      visualizationSpec={visualizationSpec}
-    />,
-    { wrapper: Provider }
-  );
-
-  await openRightSidebar();
-
-  expect(screen.queryByTestId('custom-metadata')).not.toBeInTheDocument();
-});
-
-test('hiding the controls slot removes the left sidebar even with menus configured', () => {
-  render(
-    <GeovisWorkspace
-      config={{ ...config, slots: { controls: { hidden: true } } }}
-      visualizationSpec={visualizationSpec}
-    />,
-    { wrapper: Provider }
-  );
-
-  expect(
-    screen.queryByRole('button', { name: 'Open menu' })
-  ).not.toBeInTheDocument();
-});
-
-test('closing the right sidebar brings its open button back', async () => {
-  render(
-    <GeovisWorkspace
-      config={{ ...config, legend: { description: 'Descrição' } }}
-      visualizationSpec={visualizationSpec}
-    />,
-    { wrapper: Provider }
-  );
-
-  await openRightSidebar();
-
-  await act(async () => {
-    fireEvent.click(screen.getByRole('button', { name: 'Close details' }));
-  });
-
-  expect(
-    screen.getByRole('button', { name: 'Open details' })
-  ).toBeInTheDocument();
-});
-
-test('right sidebar starts open when initialState is "open"', () => {
-  render(
-    <GeovisWorkspace
-      config={{
-        ...config,
-        legend: { description: 'Descrição' },
-        rightSidebar: { title: 'Camadas', initialState: 'open' },
-      }}
-      visualizationSpec={visualizationSpec}
-    />,
-    { wrapper: Provider }
-  );
-
-  expect(screen.getByText('Camadas')).toBeInTheDocument();
-  expect(
-    screen.queryByRole('button', { name: 'Open details' })
-  ).not.toBeInTheDocument();
-});
-
-test('getInitialSelection seeds the selection from section defaultValues', () => {
-  expect(getInitialSelection({ config })).toEqual({
-    population: undefined,
-    economy: undefined,
-  });
-
-  expect(
-    getInitialSelection({
-      config: {
-        leftSidebar: {
-          sections: [
-            {
-              id: 'economy',
-              header: { title: 'Economia' },
-              body: {
-                kind: 'variations',
-                menuId: 'economy',
-                defaultValue: 'gdp',
-                groups: [
-                  {
-                    id: 'economy',
-                    label: 'Economia',
-                    variations: [{ value: 'gdp', label: 'PIB' }],
-                  },
-                ],
-              },
-            },
-          ],
-        },
-      },
-    })
-  ).toEqual({ economy: 'gdp' });
-});
-
-test('useGeovisWorkspace throws when used outside provider', () => {
-  const BrokenComponent = () => {
-    useGeovisWorkspace();
-    return null;
-  };
-
-  expect(() => {
-    render(<BrokenComponent />, { wrapper: Provider });
-  }).toThrow(
-    'useGeovisWorkspace must be used within a GeovisWorkspaceProvider'
-  );
-});
-
-test('GeovisWorkspaceProvider exposes context to consumers', () => {
-  const Consumer = () => {
-    const { selection } = useGeovisWorkspace();
-    return <div>{selection.population ?? 'none'}</div>;
-  };
-
-  render(
-    <GeovisWorkspaceProvider config={config}>
-      <Consumer />
-    </GeovisWorkspaceProvider>,
-    { wrapper: Provider }
-  );
-
-  expect(screen.getByText('none')).toBeInTheDocument();
-});
-
-test('GeovisWorkspaceProvider manages left sidebar open state uncontrolled', async () => {
-  const Consumer = () => {
-    const { isLeftSidebarOpen, setLeftSidebarOpen } = useGeovisWorkspace();
-    return (
-      <button
-        type="button"
-        onClick={() => {
-          return setLeftSidebarOpen({ open: !isLeftSidebarOpen });
-        }}
-      >
-        {isLeftSidebarOpen ? 'open' : 'closed'}
-      </button>
-    );
-  };
-
-  render(
-    <GeovisWorkspaceProvider config={config}>
-      <Consumer />
-    </GeovisWorkspaceProvider>,
-    { wrapper: Provider }
-  );
-
-  // No controlled prop → the provider seeds and flips its own internal state.
-  expect(screen.getByRole('button', { name: 'closed' })).toBeInTheDocument();
-
-  await act(async () => {
-    fireEvent.click(screen.getByRole('button', { name: 'closed' }));
-  });
-
-  expect(screen.getByRole('button', { name: 'open' })).toBeInTheDocument();
-});
-
-test('GeovisWorkspaceProvider manages right sidebar open state uncontrolled', async () => {
-  const Consumer = () => {
-    const { isRightSidebarOpen, setRightSidebarOpen } = useGeovisWorkspace();
-    return (
-      <button
-        type="button"
-        onClick={() => {
-          return setRightSidebarOpen({ open: !isRightSidebarOpen });
-        }}
-      >
-        {isRightSidebarOpen ? 'open' : 'closed'}
-      </button>
-    );
-  };
-
-  render(
-    <GeovisWorkspaceProvider config={config}>
-      <Consumer />
-    </GeovisWorkspaceProvider>,
-    { wrapper: Provider }
-  );
-
-  // No controlled prop → the provider seeds and flips its own internal state.
-  expect(screen.getByRole('button', { name: 'closed' })).toBeInTheDocument();
-
-  await act(async () => {
-    fireEvent.click(screen.getByRole('button', { name: 'closed' }));
-  });
-
-  expect(screen.getByRole('button', { name: 'open' })).toBeInTheDocument();
-});
-
-test('warnings panel shows a resolved result warning with its i18n text and subject', async () => {
-  render(
-    <GeovisWorkspace
-      config={config}
-      visualizationSpec={visualizationSpecWithWarnings}
-    />,
-    { wrapper: Provider }
-  );
-
-  await openRightSidebar();
-
-  expect(
-    screen.getByText('This map violates the cartography policy.')
-  ).toBeInTheDocument();
-  expect(
-    screen.getByText('metadata.metricField (policy-invalid)')
-  ).toBeInTheDocument();
-});
-
-test('warnings panel shows a blocking failure that follows a successful resolve', async () => {
-  const { rerender } = render(
-    <GeovisWorkspace config={config} visualizationSpec={visualizationSpec} />,
-    { wrapper: Provider }
-  );
-
-  rerender(
-    <GeovisWorkspace
-      config={config}
-      visualizationSpec={failingVisualizationSpec}
-    />
-  );
-
-  await openRightSidebar();
-
-  expect(
-    screen.getByText('A layer references a map data id that does not exist.')
-  ).toBeInTheDocument();
-  expect(screen.getByTestId('geovis-canvas')).toBeInTheDocument();
-});
-
-test('warnings panel falls back to the raw message for a code with no catalog entry', async () => {
-  const specWithUnknownCode = {
-    ...visualizationSpec,
-    mockResult: {
-      status: 'resolved' as const,
-      spec: visualizationSpec,
-      warnings: [
-        {
-          code: 'made-up-code' as never,
-          subject: { path: '$' },
-          message: 'Raw untranslated message.',
-        },
-      ],
-    },
-  };
-
-  render(
-    <GeovisWorkspace config={config} visualizationSpec={specWithUnknownCode} />,
-    { wrapper: Provider }
-  );
-
-  await openRightSidebar();
-
-  expect(screen.getByText('Raw untranslated message.')).toBeInTheDocument();
-});
-
-test('an allowed-values repair renders one button per value and calls onRepair with the chosen value', async () => {
-  const onRepair = jest.fn();
-
-  const { rerender } = render(
-    <GeovisWorkspace
-      config={config}
-      visualizationSpec={visualizationSpec}
-      onRepair={onRepair}
-    />,
-    { wrapper: Provider }
-  );
-
-  rerender(
-    <GeovisWorkspace
-      config={config}
-      visualizationSpec={failingVisualizationSpec}
-      onRepair={onRepair}
-    />
-  );
-
-  await openRightSidebar();
-
-  const choroplethButton = screen.getByRole('button', { name: 'choropleth' });
-  const dotsButton = screen.getByRole('button', { name: 'dots' });
-  expect(choroplethButton).toBeInTheDocument();
-  expect(dotsButton).toBeInTheDocument();
-
-  await act(async () => {
-    fireEvent.click(choroplethButton);
-  });
-
-  expect(onRepair).toHaveBeenCalledWith({
-    kind: 'set-value',
-    path: 'layers[0].mapDataId',
-    value: 'choropleth',
-  });
-});
-
-test('a set-value repair calls onRepair with the option as-is', async () => {
-  const onRepair = jest.fn();
-
-  render(
-    <GeovisWorkspace
-      config={config}
-      visualizationSpec={visualizationSpecWithWarnings}
-      onRepair={onRepair}
-    />,
-    { wrapper: Provider }
-  );
-
-  await openRightSidebar();
-
-  await act(async () => {
-    fireEvent.click(screen.getByRole('button', { name: "Use 'safe-field'" }));
-  });
-
-  expect(onRepair).toHaveBeenCalledWith({
-    kind: 'set-value',
-    path: 'metadata.metricField',
-    value: 'safe-field',
-    label: "Use 'safe-field'",
-  });
-});
-
-test('a set-value repair without a label falls back to the stringified value', async () => {
-  const specWithUnlabeledRepair = {
-    ...visualizationSpec,
-    mockResult: {
-      status: 'resolved' as const,
-      spec: visualizationSpec,
-      warnings: [
-        {
-          code: 'policy-violation' as const,
-          subject: { path: 'metadata.metricField' },
-          message: 'Spec violates policy.',
-          repair: [
-            {
-              kind: 'set-value' as const,
-              path: 'metadata.metricField',
-              value: 'safe-field',
-            },
-          ],
-        },
-      ],
-    },
-  };
-
-  render(
-    <GeovisWorkspace
-      config={config}
-      visualizationSpec={specWithUnlabeledRepair}
-    />,
-    { wrapper: Provider }
-  );
-
-  await openRightSidebar();
-
-  expect(
-    screen.getByRole('button', { name: 'safe-field' })
-  ).toBeInTheDocument();
-});
-
-test('repair buttons render disabled rather than hidden when onRepair is omitted', async () => {
-  render(
-    <GeovisWorkspace
-      config={config}
-      visualizationSpec={visualizationSpecWithWarnings}
-    />,
-    { wrapper: Provider }
-  );
-
-  await openRightSidebar();
-
-  expect(
-    screen.getByRole('button', { name: "Use 'safe-field'" })
-  ).toBeDisabled();
-});
-
-test('cold start: a failing result before any resolve shows the empty state instead of the canvas', () => {
-  render(
-    <GeovisWorkspace
-      config={config}
-      visualizationSpec={failingVisualizationSpec}
-    />,
-    { wrapper: Provider }
-  );
-
-  expect(screen.getByText('Map could not be shown')).toBeInTheDocument();
-  expect(
-    screen.getByText('A layer references a map data id that does not exist.')
-  ).toBeInTheDocument();
-  expect(screen.queryByTestId('geovis-canvas')).not.toBeInTheDocument();
-});
-
-test('cold start: the warnings panel itself suppresses issues even when the sidebar shows for the legend slot', async () => {
-  render(
-    <GeovisWorkspace
-      config={{ ...config, legend: { description: 'Descrição' } }}
-      visualizationSpec={failingVisualizationSpec}
-    />,
-    { wrapper: Provider }
-  );
-
-  await openRightSidebar();
-
-  expect(screen.getByText('Descrição')).toBeInTheDocument();
-  // Shown once, by the map's cold-start empty state — not a second time by
-  // the warnings panel, which is also visible now (for the legend slot).
-  expect(
-    screen.getAllByText('A layer references a map data id that does not exist.')
-  ).toHaveLength(1);
-});
-
-test('once a resolve succeeds, a later failure keeps the canvas instead of re-showing the cold-start state', () => {
-  const { rerender } = render(
-    <GeovisWorkspace
-      config={config}
-      visualizationSpec={failingVisualizationSpec}
-    />,
-    { wrapper: Provider }
-  );
-
-  expect(screen.getByText('Map could not be shown')).toBeInTheDocument();
-
-  rerender(
-    <GeovisWorkspace config={config} visualizationSpec={visualizationSpec} />
-  );
-
-  expect(screen.getByTestId('geovis-canvas')).toBeInTheDocument();
-  expect(screen.queryByText('Map could not be shown')).not.toBeInTheDocument();
-
-  rerender(
-    <GeovisWorkspace
-      config={config}
-      visualizationSpec={failingVisualizationSpec}
-    />
-  );
-
-  expect(screen.getByTestId('geovis-canvas')).toBeInTheDocument();
-  expect(screen.queryByText('Map could not be shown')).not.toBeInTheDocument();
-});
-
-test('right sidebar is absent when nothing is selected on the map', () => {
-  render(
-    <GeovisWorkspace config={config} visualizationSpec={visualizationSpec} />,
-    { wrapper: Provider }
-  );
-
-  expect(
-    screen.queryByRole('button', { name: 'Open details' })
-  ).not.toBeInTheDocument();
-});
-
-test('inspector panel shows the selected feature and clearing the selection removes it', async () => {
-  const { rerender } = render(
-    <GeovisWorkspace config={config} visualizationSpec={visualizationSpec} />,
-    { wrapper: Provider }
-  );
-
-  expect(
-    screen.queryByRole('button', { name: 'Open details' })
-  ).not.toBeInTheDocument();
-
-  rerender(
-    <GeovisWorkspace
-      config={config}
-      visualizationSpec={{
-        ...visualizationSpec,
-        mockClick: { layerId: 'districts-fill', featureId: 'sp', value: 42 },
-      }}
-    />
-  );
-
-  await openRightSidebar();
-
-  expect(screen.getByText('districts-fill')).toBeInTheDocument();
-  expect(screen.getByText('42')).toBeInTheDocument();
-  expect(screen.getByText('sp')).toBeInTheDocument();
-});
-
-test('inspector panel shows a fallback when the selected feature has no value', async () => {
-  render(
-    <GeovisWorkspace
-      config={config}
-      visualizationSpec={{
-        ...visualizationSpec,
-        mockClick: { layerId: 'districts-fill', featureId: 'sp', value: null },
-      }}
-    />,
-    { wrapper: Provider }
-  );
-
-  await openRightSidebar();
-
-  expect(screen.getByText('No value')).toBeInTheDocument();
-});
-
-test('dismissing the inspector selection clears the panel and closes the sidebar content', async () => {
-  render(
-    <GeovisWorkspace
-      config={config}
-      visualizationSpec={{
-        ...visualizationSpec,
-        mockClick: { layerId: 'districts-fill', featureId: 'sp', value: 42 },
-      }}
-    />,
-    { wrapper: Provider }
-  );
-
-  await openRightSidebar();
-
-  expect(screen.getByText('districts-fill')).toBeInTheDocument();
-
-  await act(async () => {
-    fireEvent.click(screen.getByRole('button', { name: 'Dismiss selection' }));
-  });
-
-  expect(screen.queryByText('districts-fill')).not.toBeInTheDocument();
-});
-
-test('metadata panel shows the map type and pluralized source count', async () => {
-  const specWithMetadata = {
-    ...visualizationSpec,
-    mapType: 'choropleth',
-    sources: [{ id: 's1' }, { id: 's2' }],
-  };
-
-  render(
-    <GeovisWorkspace config={config} visualizationSpec={specWithMetadata} />,
-    { wrapper: Provider }
-  );
-
-  await openRightSidebar();
-
-  expect(screen.getByText('Map type: choropleth')).toBeInTheDocument();
-  expect(screen.getByText('2 sources')).toBeInTheDocument();
-});
-
-test('metadata panel uses the singular form for a single source', async () => {
-  const specWithOneSource = {
-    ...visualizationSpec,
-    sources: [{ id: 's1' }],
-  };
-
-  render(
-    <GeovisWorkspace config={config} visualizationSpec={specWithOneSource} />,
-    { wrapper: Provider }
-  );
-
-  await openRightSidebar();
-
-  expect(screen.getByText('1 source')).toBeInTheDocument();
-});
-
-test('metadata panel contributes no content and no sidebar when the spec has neither mapType nor sources', () => {
-  render(
-    <GeovisWorkspace config={config} visualizationSpec={visualizationSpec} />,
-    { wrapper: Provider }
-  );
-
-  expect(
-    screen.queryByRole('button', { name: 'Open details' })
-  ).not.toBeInTheDocument();
-});
-
-test('metadata panel is absent while another slot keeps the sidebar open', async () => {
-  render(
-    <GeovisWorkspace
-      config={{ ...config, legend: { description: 'Descrição' } }}
-      visualizationSpec={visualizationSpec}
-    />,
-    { wrapper: Provider }
-  );
-
-  await openRightSidebar();
-
-  expect(screen.getByText('Descrição')).toBeInTheDocument();
-  expect(screen.queryByText(/Map type:/)).not.toBeInTheDocument();
-  expect(screen.queryByText(/source/)).not.toBeInTheDocument();
-});
-
-test('right sidebar detail API fetches on click, auto-opens, and renders the detail', async () => {
-  const onFeatureSelect = jest.fn(() => {
-    return Promise.resolve({ name: 'Cozinha X' });
-  });
-
-  render(
-    <GeovisWorkspace
-      config={{
-        rightSidebar: {
-          title: 'Detalhe',
-          onFeatureSelect,
-          renderDetails: ({ loading, data }) => {
-            if (loading) return <span>Carregando</span>;
-            return <span>{(data as { name: string } | null)?.name}</span>;
-          },
-        },
-      }}
-      visualizationSpec={{
-        ...visualizationSpec,
-        mockClick: { layerId: 'kitchens', featureId: 'k1', value: null },
-      }}
-    />,
-    { wrapper: Provider }
-  );
-
-  // The detail renders without a manual open — an accepted click auto-opens the
-  // sidebar, so its reopen button is gone.
-  expect(await screen.findByText('Cozinha X')).toBeInTheDocument();
-  expect(
-    screen.queryByRole('button', { name: 'Open details' })
-  ).not.toBeInTheDocument();
-  expect(onFeatureSelect).toHaveBeenCalledWith(
-    expect.objectContaining({ layerId: 'kitchens', featureId: 'k1' })
-  );
-});
-
-test('right sidebar closes when the selection clears (click outside), not only via the close button', async () => {
-  const onFeatureSelect = jest.fn(() => {
-    return Promise.resolve({ name: 'Cozinha X' });
-  });
-
-  // `legend.description` keeps the right sidebar populated after the selection
-  // clears, so its open state stays observable via the reopen button instead of
-  // being hidden for lack of content.
-  const configWithDetail: GeovisWorkspaceConfig = {
-    legend: { description: 'Descrição da legenda.' },
-    rightSidebar: {
-      onFeatureSelect,
-      renderDetails: ({ loading, data }) => {
-        if (loading) return <span>Carregando</span>;
-        return <span>{(data as { name: string } | null)?.name}</span>;
-      },
-    },
-  };
-
-  const { rerender } = render(
-    <GeovisWorkspace
-      config={configWithDetail}
-      visualizationSpec={{
-        ...visualizationSpec,
-        mockClick: { layerId: 'kitchens', featureId: 'k1', value: null },
-      }}
-    />,
-    { wrapper: Provider }
-  );
-
-  // Clicking a feature auto-opens the sidebar: the detail shows and the reopen
-  // button is gone.
-  expect(await screen.findByText('Cozinha X')).toBeInTheDocument();
-  expect(
-    screen.queryByRole('button', { name: 'Open details' })
-  ).not.toBeInTheDocument();
-
-  // Clicking outside clears the selection (no manual close-button press).
-  rerender(
-    <GeovisWorkspace
-      config={configWithDetail}
-      visualizationSpec={{ ...visualizationSpec, mockClick: null }}
-    />
-  );
-
-  await act(async () => {});
-
-  // The detail is gone and the sidebar is closed — its reopen button is back,
-  // even though the legend description still keeps the sidebar populated. Before
-  // the fix, the open state stayed `true` here and a pushed-aside legend would
-  // never return to rest.
-  expect(screen.queryByText('Cozinha X')).not.toBeInTheDocument();
-  expect(
-    screen.getByRole('button', { name: 'Open details' })
-  ).toBeInTheDocument();
-});
-
-test('right sidebar detail API ignores clicks rejected by shouldOpen', async () => {
-  const onFeatureSelect = jest.fn(() => {
-    return Promise.resolve({ name: 'X' });
-  });
-
-  render(
-    <GeovisWorkspace
-      config={{
-        rightSidebar: {
-          onFeatureSelect,
-          shouldOpen: (info) => {
-            return info.layerId === 'kitchens';
-          },
-          renderDetails: ({ data }) => {
-            return (
-              <span>{(data as { name: string } | null)?.name ?? 'none'}</span>
-            );
-          },
-        },
-      }}
-      visualizationSpec={{
-        ...visualizationSpec,
-        mockClick: { layerId: 'other', featureId: 'x', value: null },
-      }}
-    />,
-    { wrapper: Provider }
-  );
-
-  await act(async () => {});
-
-  expect(
-    screen.queryByRole('button', { name: 'Open details' })
-  ).not.toBeInTheDocument();
-  expect(onFeatureSelect).not.toHaveBeenCalled();
-});
-
-test('right sidebar detail API surfaces onFeatureSelect errors to renderDetails', async () => {
-  const onFeatureSelect = jest.fn(() => {
-    return Promise.reject(new Error('boom'));
-  });
-
-  render(
-    <GeovisWorkspace
-      config={{
-        rightSidebar: {
-          onFeatureSelect,
-          shouldOpen: () => {
-            return true;
-          },
-          renderDetails: ({ loading, error }) => {
-            if (loading) return <span>Carregando</span>;
-            return error ? <span>Erro</span> : null;
-          },
-        },
-      }}
-      visualizationSpec={{
-        ...visualizationSpec,
-        mockClick: { layerId: 'kitchens', featureId: 'k1', value: null },
-      }}
-    />,
-    { wrapper: Provider }
-  );
-
-  expect(await screen.findByText('Erro')).toBeInTheDocument();
-});
-
-test('right sidebar detail API runs onFeatureSelect even without renderDetails', async () => {
-  const onFeatureSelect = jest.fn(() => {
-    return Promise.resolve(null);
-  });
-
-  render(
-    <GeovisWorkspace
-      config={{ rightSidebar: { onFeatureSelect } }}
-      visualizationSpec={{
-        ...visualizationSpec,
-        mockClick: { layerId: 'kitchens', featureId: 'k1', value: null },
-      }}
-    />,
-    { wrapper: Provider }
-  );
-
-  await act(async () => {});
-
-  expect(onFeatureSelect).toHaveBeenCalledWith(
-    expect.objectContaining({ layerId: 'kitchens' })
-  );
-});
-
-test('right sidebar shows inspector content for renderDetails without onFeatureSelect', async () => {
-  render(
-    <GeovisWorkspace
-      config={{
-        rightSidebar: {
-          renderDetails: ({ data }) => {
-            return <span>{String(data)}</span>;
-          },
-        },
-      }}
-      visualizationSpec={{
-        ...visualizationSpec,
-        mockClick: { layerId: 'kitchens', featureId: 'k1', value: null },
-      }}
-    />,
-    { wrapper: Provider }
-  );
-
-  await act(async () => {});
-
-  // With no `onFeatureSelect`, an accepted click gives the inspector slot
-  // content (so the sidebar exists) but nothing is fetched or auto-opened.
-  expect(
-    screen.getByRole('button', { name: 'Open details' })
-  ).toBeInTheDocument();
-});
-
-test('right sidebar detail API does not fetch until a feature is clicked', async () => {
-  const onFeatureSelect = jest.fn(() => {
-    return Promise.resolve({ name: 'X' });
-  });
-
-  render(
-    <GeovisWorkspace
-      config={{
-        legend: { description: 'Descrição' },
-        rightSidebar: {
-          onFeatureSelect,
-          renderDetails: ({ data }) => {
-            return (
-              <span>{(data as { name: string } | null)?.name ?? '—'}</span>
-            );
-          },
-        },
-      }}
-      visualizationSpec={visualizationSpec}
-    />,
-    { wrapper: Provider }
-  );
-
-  await act(async () => {});
-  await openRightSidebar();
-
-  expect(screen.getByText('Descrição')).toBeInTheDocument();
-  expect(onFeatureSelect).not.toHaveBeenCalled();
-});
-
-test('right sidebar detail API ignores a stale fetch when the click changes', async () => {
-  let resolveFirst: (value: unknown) => void = () => {};
-
-  const onFeatureSelect = jest.fn((info: { featureId: string | number }) => {
-    if (info.featureId === 'k1') {
-      return new Promise<unknown>((resolve) => {
-        resolveFirst = resolve;
-      });
-    }
-    return Promise.resolve({ name: 'Second' });
-  });
-
-  const detailConfig = {
-    rightSidebar: {
-      onFeatureSelect,
-      renderDetails: ({ data }: { data: unknown }) => {
-        return (
-          <span>{(data as { name: string } | null)?.name ?? 'pending'}</span>
-        );
-      },
-    },
-  };
-
-  const { rerender } = render(
-    <GeovisWorkspace
-      config={detailConfig}
-      visualizationSpec={{
-        ...visualizationSpec,
-        mockClick: { layerId: 'kitchens', featureId: 'k1', value: null },
-      }}
-    />,
-    { wrapper: Provider }
-  );
-
-  rerender(
-    <GeovisWorkspace
-      config={detailConfig}
-      visualizationSpec={{
-        ...visualizationSpec,
-        mockClick: { layerId: 'kitchens', featureId: 'k2', value: null },
-      }}
-    />
-  );
-
-  // Resolve the now-stale first request after the selection already moved on.
-  await act(async () => {
-    resolveFirst({ name: 'First' });
-  });
-
-  expect(await screen.findByText('Second')).toBeInTheDocument();
-  expect(screen.queryByText('First')).not.toBeInTheDocument();
-});
-
-test('right sidebar detail API skips the fetch for a rejected click while the sidebar stays up', async () => {
-  const onFeatureSelect = jest.fn(() => {
-    return Promise.resolve({ name: 'X' });
-  });
-
-  render(
-    <GeovisWorkspace
-      config={{
-        legend: { description: 'Descrição' },
-        rightSidebar: {
-          onFeatureSelect,
-          shouldOpen: (info) => {
-            return info.layerId === 'kitchens';
-          },
-          renderDetails: ({ data }) => {
-            return (
-              <span>{(data as { name: string } | null)?.name ?? '—'}</span>
-            );
-          },
-        },
-      }}
-      visualizationSpec={{
-        ...visualizationSpec,
-        mockClick: { layerId: 'other', featureId: 'x', value: null },
-      }}
-    />,
-    { wrapper: Provider }
-  );
-
-  await act(async () => {});
-  await openRightSidebar();
-
-  // The legend keeps the sidebar mounted, but the rejected click drives no fetch.
-  expect(screen.getByText('Descrição')).toBeInTheDocument();
-  expect(onFeatureSelect).not.toHaveBeenCalled();
-});
-
-test('right sidebar detail API shows loading again when a new click supersedes a resolved one', async () => {
-  let resolveSecond: (value: unknown) => void = () => {};
-
-  const onFeatureSelect = jest.fn((info: { featureId: string | number }) => {
-    if (info.featureId === 'k1') {
-      return Promise.resolve({ name: 'First' });
-    }
-    return new Promise<unknown>((resolve) => {
-      resolveSecond = resolve;
-    });
-  });
-
-  const detailConfig = {
-    rightSidebar: {
-      onFeatureSelect,
-      renderDetails: ({
-        loading,
-        data,
-      }: {
-        loading: boolean;
-        data: unknown;
-      }) => {
-        if (loading) return <span>Carregando</span>;
-        return <span>{(data as { name: string } | null)?.name}</span>;
-      },
-    },
-  };
-
-  const { rerender } = render(
-    <GeovisWorkspace
-      config={detailConfig}
-      visualizationSpec={{
-        ...visualizationSpec,
-        mockClick: { layerId: 'k', featureId: 'k1', value: null },
-      }}
-    />,
-    { wrapper: Provider }
-  );
-
-  expect(await screen.findByText('First')).toBeInTheDocument();
-
-  rerender(
-    <GeovisWorkspace
-      config={detailConfig}
-      visualizationSpec={{
-        ...visualizationSpec,
-        mockClick: { layerId: 'k', featureId: 'k2', value: null },
-      }}
-    />
-  );
-
-  // The resolved k1 detail must not linger while k2's fetch is still pending.
-  expect(await screen.findByText('Carregando')).toBeInTheDocument();
-
-  await act(async () => {
-    resolveSecond({ name: 'Second' });
-  });
-
-  expect(await screen.findByText('Second')).toBeInTheDocument();
 });
