+++ packages/geovis/tests/unit/tests/adapters/maplibre/MapLibreAdapter.test.ts
@@ -1,21 +1,22 @@
 /**
- * Tests for MapLibreAdapter factory isolation.
- * Verifies that each adapter instance has independent state
- * and that destroying one does not affect another.
+ * Tests for MapLibreAdapter factory isolation and runtime behaviour: instance
+ * independence, selection, view sync, patch application and basemap labels.
+ *
+ * Spec translation lives in MapLibreAdapter.translation.test.ts and
+ * feature-state handling in MapLibreAdapter.mapData.test.ts.
  */
 
 import maplibregl from 'maplibre-gl';
-import createMapLibreAdapter, {
-  toMaplibreLayer,
-  toMaplibreSource,
-} from 'src/adapters/maplibre/MapLibreAdapter';
-import type {
-  DataSource,
-  HeatmapPaint,
-  RasterPaint,
-  SymbolPaint,
-  VisualizationLayer,
-} from 'src/spec/types';
+import createMapLibreAdapter from 'src/adapters/maplibre/MapLibreAdapter';
+import type { VisualizationLayer } from 'src/spec/types';
+
+import {
+  installMapLibreDomMocks,
+  makeContainer,
+  makeMapMock,
+  makeSpec,
+  resetMapLibreDomMocks,
+} from './mapLibreAdapterTestUtils';
 
 jest.mock('maplibre-gl', () => {
   return {
@@ -26,85 +27,10 @@ jest.mock('maplibre-gl', () => {
   };
 });
 
-// Mock DOM APIs used by injectMaplibreCSS
-Object.defineProperty(global, 'document', {
-  value: {
-    getElementById: jest.fn(() => {
-      return null;
-    }),
-    createElement: jest.fn(() => {
-      return { id: '', rel: '', href: '', style: {} };
-    }),
-    head: { appendChild: jest.fn() },
-  },
-  writable: true,
-});
-
-Object.defineProperty(global, 'URL', {
-  value: jest.fn(() => {
-    return { href: 'mocked-css-url' };
-  }),
-  writable: true,
-});
-
-const makeMapMock = () => {
-  return {
-    on: jest.fn(),
-    once: jest.fn(),
-    remove: jest.fn(),
-    addControl: jest.fn(),
-    addSource: jest.fn(),
-    addLayer: jest.fn(),
-    getSource: jest.fn(() => {
-      return null;
-    }),
-    getLayer: jest.fn(() => {
-      return null;
-    }),
-    removeLayer: jest.fn(),
-    removeSource: jest.fn(),
-    moveLayer: jest.fn(),
-    getStyle: jest.fn(() => {
-      return { layers: [] };
-    }),
-    isStyleLoaded: jest.fn(() => {
-      return true;
-    }),
-    setLayoutProperty: jest.fn(),
-    setPaintProperty: jest.fn(),
-    setFeatureState: jest.fn(),
-    setFilter: jest.fn(),
-    setStyle: jest.fn(),
-    setCenter: jest.fn(),
-    setZoom: jest.fn(),
-    setMaxZoom: jest.fn(),
-    setMinZoom: jest.fn(),
-    setPitch: jest.fn(),
-    setBearing: jest.fn(),
-  };
-};
-
-const makeSpec = () => {
-  return {
-    engine: 'maplibre' as const,
-    view: { center: [-46.6, -23.5] as [number, number], zoom: 10 },
-    sources: [],
-    layers: [],
-  };
-};
-
-const makeContainer = () => {
-  return { style: {} } as unknown as HTMLElement;
-};
+installMapLibreDomMocks();
 
 beforeEach(() => {
-  jest.clearAllMocks();
-  (document.getElementById as jest.Mock).mockReturnValue(null);
-  (document.createElement as jest.Mock).mockReturnValue({
-    id: '',
-    rel: '',
-    href: '',
-  });
+  resetMapLibreDomMocks();
 });
 
 describe('createMapLibreAdapter', () => {
@@ -349,113 +275,6 @@ describe('createMapLibreAdapter — setSelection (PRD-002 select-feature)', () =
   });
 });
 
-describe('toMaplibreSource', () => {
-  test('geojson source with url string', () => {
-    const source: DataSource = {
-      id: 's1',
-      type: 'geojson',
-      data: 'https://example.com/data.geojson',
-    };
-    expect(toMaplibreSource(source)).toMatchObject({
-      type: 'geojson',
-      data: 'https://example.com/data.geojson',
-    });
-  });
-
-  test('vector-tiles source', () => {
-    const source: DataSource = {
-      id: 's2',
-      type: 'vector-tiles',
-      tiles: ['https://example.com/{z}/{x}/{y}.pbf'],
-      minzoom: 0,
-      maxzoom: 14,
-      attribution: '© Example',
-    };
-    expect(toMaplibreSource(source)).toMatchObject({
-      type: 'vector',
-      tiles: ['https://example.com/{z}/{x}/{y}.pbf'],
-      minzoom: 0,
-      maxzoom: 14,
-      attribution: '© Example',
-    });
-  });
-
-  test('raster-tiles source applies default tileSize 256', () => {
-    const source: DataSource = {
-      id: 's3',
-      type: 'raster-tiles',
-      tiles: ['https://example.com/{z}/{x}/{y}.png'],
-    };
-    expect(toMaplibreSource(source)).toMatchObject({
-      type: 'raster',
-      tileSize: 256,
-    });
-  });
-
-  test('raster-tiles source respects explicit tileSize 512', () => {
-    const source: DataSource = {
-      id: 's3b',
-      type: 'raster-tiles',
-      tiles: ['https://tile.example/{z}/{x}/{y}.png'],
-      tileSize: 512,
-    };
-    expect(toMaplibreSource(source)).toMatchObject({
-      type: 'raster',
-      tileSize: 512,
-    });
-  });
-
-  test('image source', () => {
-    const source: DataSource = {
-      id: 's4',
-      type: 'image',
-      url: 'https://example.com/image.png',
-      coordinates: [
-        [-80, 25],
-        [-80, 26],
-        [-79, 26],
-        [-79, 25],
-      ],
-    };
-    expect(toMaplibreSource(source)).toMatchObject({
-      type: 'image',
-      url: 'https://example.com/image.png',
-    });
-  });
-
-  test('raster-dem source with encoding', () => {
-    const source: DataSource = {
-      id: 's5',
-      type: 'raster-dem',
-      tiles: ['https://tiles.example/{z}/{x}/{y}.png'],
-      encoding: 'terrarium',
-    };
-    expect(toMaplibreSource(source)).toMatchObject({
-      type: 'raster-dem',
-      encoding: 'terrarium',
-      tileSize: 256,
-    });
-  });
-
-  test('video source', () => {
-    const source: DataSource = {
-      id: 's6',
-      type: 'video',
-      urls: ['https://example.com/v.mp4'],
-      coordinates: [
-        [-122, 37],
-        [-122, 38],
-        [-121, 38],
-        [-121, 37],
-      ],
-    };
-    expect(toMaplibreSource(source)).toMatchObject({
-      type: 'video',
-      urls: ['https://example.com/v.mp4'],
-    });
-  });
-});
-
 describe('applyPatch — camelCase to MapLibre key translation', () => {
   const mountAdapter = () => {
     const map = makeMapMock();
@@ -645,209 +464,6 @@ describe('update() — view state sync', () => {
   });
 });
 
-describe('toMaplibreLayer', () => {
-  const base = { id: 'l1', sourceId: 's1', visible: true as const };
-
-  test('polygon → fill with defaults', () => {
-    const layer: VisualizationLayer = { ...base, geometry: 'polygon' };
-    expect(toMaplibreLayer(layer)).toMatchObject({
-      type: 'fill',
-      paint: { 'fill-color': '#3b82f6' },
-    });
-  });
-
-  test('polygon with custom paint', () => {
-    const layer: VisualizationLayer = {
-      ...base,
-      geometry: 'polygon',
-      paint: { fillColor: '#ff0000', lineColor: '#000' },
-    };
-    expect(toMaplibreLayer(layer)).toMatchObject({
-      paint: { 'fill-color': '#ff0000' },
-    });
-  });
-
-  test('polygon uses active categorical legend expression as fill-color', () => {
-    const layer: VisualizationLayer = {
-      ...base,
-      geometry: 'polygon',
-      paint: { fillColor: '#ff0000' },
-      legends: [
-        {
-          id: 'status',
-          colorBy: {
-            type: 'categorical',
-            property: 'status',
-            mapping: {
-              open: '#16a34a',
-              closed: '#dc2626',
-            },
-            defaultColor: '#6b7280',
-          },
-        },
-      ],
-      activeLegendId: 'status',
-    };
-
-    expect(toMaplibreLayer(layer)).toMatchObject({
-      paint: {
-        'fill-color': [
-          'match',
-          [
-            'to-string',
-            ['coalesce', ['feature-state', 'value'], '__missing__'],
-          ],
-          'open',
-          '#16a34a',
-          'closed',
-          '#dc2626',
-          '#6b7280',
-        ],
-      },
-    });
-  });
-
-  test('polygon uses threshold breaks when active legend scale is threshold', () => {
-    const layer: VisualizationLayer = {
-      ...base,
-      geometry: 'polygon',
-      legends: [
-        {
-          id: 'population',
-          colorBy: {
-            type: 'quantitative',
-            property: 'population',
-            scale: 'threshold',
-            thresholds: [30, 10, 10],
-            colors: ['#eff6ff', '#bfdbfe', '#60a5fa'],
-            defaultColor: '#1e3a8a',
-          },
-        },
-      ],
-      activeLegendId: 'population',
-    };
-
-    expect(toMaplibreLayer(layer)).toMatchObject({
-      paint: {
-        'fill-color': [
-          'step',
-          ['to-number', ['coalesce', ['feature-state', 'value'], 0], 0],
-          '#1e3a8a',
-          10,
-          '#bfdbfe',
-          30,
-          '#60a5fa',
-        ],
-      },
-    });
-  });
-
-  test('polygon keeps static fill-color when active legend id does not resolve', () => {
-    const layer: VisualizationLayer = {
-      ...base,
-      geometry: 'polygon',
-      paint: { fillColor: '#ff0000' },
-      legends: [
-        {
-          id: 'status',
-          colorBy: {
-            type: 'categorical',
-            property: 'status',
-            mapping: { open: '#16a34a' },
-          },
-        },
-      ],
-      activeLegendId: 'missing-id',
-    };
-
-    expect(toMaplibreLayer(layer)).toMatchObject({
-      paint: { 'fill-color': '#ff0000' },
-    });
-  });
-
-  test('line → line with defaults', () => {
-    const layer: VisualizationLayer = { ...base, geometry: 'line' };
-    expect(toMaplibreLayer(layer)).toMatchObject({
-      type: 'line',
-      paint: { 'line-color': '#3b82f6', 'line-width': 2 },
-    });
-  });
-
-  test('point → circle', () => {
-    const layer: VisualizationLayer = { ...base, geometry: 'point' };
-    expect(toMaplibreLayer(layer)).toMatchObject({ type: 'circle' });
-  });
-
-  test('symbol → symbol layer with text/icon paint and layout', () => {
-    const layer: VisualizationLayer = {
-      ...base,
-      geometry: 'symbol',
-      paint: { textField: 'Hello', textSize: 14 } as SymbolPaint,
-    };
-    const result = toMaplibreLayer(layer);
-    expect(result).toMatchObject({ type: 'symbol' });
-    expect(
-      (result as { layout: Record<string, unknown> }).layout
-    ).toMatchObject({ 'text-field': 'Hello', 'text-size': 14 });
-    expect((result as { paint: Record<string, unknown> }).paint).toMatchObject({
-      'text-color': '#000000',
-    });
-  });
-
-  test('heatmap → heatmap layer with heatmap paint properties', () => {
-    const layer: VisualizationLayer = {
-      ...base,
-      geometry: 'heatmap',
-      paint: { heatmapRadius: 20 } as HeatmapPaint,
-    };
-    const result = toMaplibreLayer(layer);
-    expect(result).toMatchObject({ type: 'heatmap' });
-    expect((result as { paint: Record<string, unknown> }).paint).toMatchObject({
-      'heatmap-radius': 20,
-      'heatmap-intensity': 1,
-      'heatmap-weight': 1,
-    });
-  });
-
-  test('raster → raster layer with raster paint properties', () => {
-    const layer: VisualizationLayer = {
-      ...base,
-      geometry: 'raster',
-      paint: { rasterOpacity: 0.5 } as RasterPaint,
-    };
-    const result = toMaplibreLayer(layer);
-    expect(result).toMatchObject({ type: 'raster' });
-    expect((result as { paint: Record<string, unknown> }).paint).toMatchObject({
-      'raster-opacity': 0.5,
-    });
-  });
-
-  test('raster paint defaults raster-opacity to 1 when unset', () => {
-    const layer: VisualizationLayer = { ...base, geometry: 'raster' };
-    const result = toMaplibreLayer(layer);
-    expect((result as { paint: Record<string, unknown> }).paint).toMatchObject({
-      'raster-opacity': 1,
-    });
-  });
-
-  test('visible: false → layout visibility none', () => {
-    const layer: VisualizationLayer = {
-      ...base,
-      geometry: 'polygon',
-      visible: false,
-    };
-    expect(toMaplibreLayer(layer)).toHaveProperty('layout.visibility', 'none');
-  });
-
-  test('visible: true → layout visibility visible', () => {
-    const layer: VisualizationLayer = { ...base, geometry: 'line' };
-    expect(toMaplibreLayer(layer)).toHaveProperty(
-      'layout.visibility',
-      'visible'
-    );
-  });
-});
-
 describe('applyPatch — add / remove operations', () => {
   const mountAdapter = () => {
     const map = makeMapMock();
@@ -941,735 +557,6 @@ describe('syncSourcesAndLayers — GeoJSON setData', () => {
     adapter.update({ ...initialSpec });
 
     expect(setData).not.toHaveBeenCalled();
-  });
-});
-
-describe('mapData — feature-state application', () => {
-  const makeMapWithEvents = () => {
-    const handlers: Record<string, Array<(...a: unknown[]) => void>> = {};
-    const map = {
-      ...makeMapMock(),
-      on: jest.fn((evt: string, cb: (...a: unknown[]) => void) => {
-        handlers[evt] = handlers[evt] ?? [];
-        handlers[evt].push(cb);
-      }),
-      off: jest.fn((evt: string, cb: (...a: unknown[]) => void) => {
-        handlers[evt] = (handlers[evt] ?? []).filter((h) => {
-          return h !== cb;
-        });
-      }),
-      // Source treated as already registered so applyMapDataToSource proceeds.
-      getSource: jest.fn(() => {
-        return { setData: jest.fn() } as unknown as ReturnType<
-          typeof maplibregl.Map.prototype.getSource
-        >;
-      }),
-      isSourceLoaded: jest.fn(() => {
-        return true;
-      }),
-      setFeatureState: jest.fn(),
-      removeFeatureState: jest.fn(),
-      querySourceFeatures: jest.fn(() => {
-        return [];
-      }),
-    };
-    const fire = (evt: string, ...args: unknown[]) => {
-      for (const cb of handlers[evt] ?? []) cb(...args);
-    };
-    return { map, fire };
-  };
-
-  const baseSpec = (mapData?: unknown) => {
-    return {
-      ...makeSpec(),
-      sources: [
-        {
-          id: 'states',
-          type: 'geojson' as const,
-          data: { type: 'FeatureCollection' as const, features: [] },
-        },
-      ],
-      layers: [
-        { id: 'fill', sourceId: 'states', geometry: 'polygon' as const },
-      ],
-      ...(mapData ? { mapData } : {}),
-    };
-  };
-
-  // 3.1
-  test('applies setFeatureState for each row after load (no joinKey → uses feature.id)', () => {
-    const { map, fire } = makeMapWithEvents();
-    jest.mocked(maplibregl.Map).mockImplementationOnce(() => {
-      return map as never;
-    });
-    const adapter = createMapLibreAdapter();
-    adapter.mount(
-      makeContainer(),
-      baseSpec([
-        {
-          mapDataId: 'pop',
-          mapId: 'states',
-          data: [
-            { geometryId: 'BR', value: 211 },
-            { geometryId: 'AR', value: 45 },
-          ],
-        },
-      ]),
-      'v'
-    );
-
-    fire('load');
-
-    expect(map.setFeatureState).toHaveBeenCalledTimes(2);
-    expect(map.setFeatureState).toHaveBeenCalledWith(
-      { source: 'states', id: 'BR' },
-      { value: 211 }
-    );
-    expect(map.setFeatureState).toHaveBeenCalledWith(
-      { source: 'states', id: 'AR' },
-      { value: 45 }
-    );
-  });
-
-  // 3.2 — numeric geometryId stored as string is coerced to number for setFeatureState
-  test('coerces string geometryId to number when it represents a finite number', () => {
-    const { map, fire } = makeMapWithEvents();
-    jest.mocked(maplibregl.Map).mockImplementationOnce(() => {
-      return map as never;
-    });
-    const adapter = createMapLibreAdapter();
-    adapter.mount(
-      makeContainer(),
-      baseSpec([
-        {
-          mapDataId: 'pop',
-          mapId: 'states',
-          data: [
-            { geometryId: '1', value: 100 },
-            { geometryId: '2', value: 200 },
-          ],
-        },
-      ]),
-      'v'
-    );
-
-    fire('load');
-
-    expect(map.setFeatureState).toHaveBeenCalledTimes(2);
-    expect(map.setFeatureState).toHaveBeenCalledWith(
-      { source: 'states', id: 1 },
-      { value: 100 }
-    );
-    expect(map.setFeatureState).toHaveBeenCalledWith(
-      { source: 'states', id: 2 },
-      { value: 200 }
-    );
-  });
-
-  // 3.2b — removeFeatureState also coerces string geometryId to number
-  test('op:remove coerces string geometryId to number for removeFeatureState', () => {
-    const { map, fire } = makeMapWithEvents();
-    jest.mocked(maplibregl.Map).mockImplementationOnce(() => {
-      return map as never;
-    });
-    const adapter = createMapLibreAdapter();
-    adapter.mount(
-      makeContainer(),
-      baseSpec([
-        {
-          mapDataId: 'pop',
-          mapId: 'states',
-          data: [{ geometryId: '1', value: 100 }],
-        },
-      ]),
-      'v'
-    );
-    fire('load');
-
-    adapter.applyPatch?.({ target: 'mapData', op: 'remove', value: 'pop' });
-
-    expect(map.removeFeatureState).toHaveBeenCalledWith(
-      {
-        source: 'states',
-        id: 1,
-      },
-      'value'
-    );
-  });
-
-  // 3.3
-  test('applyPatch granular replace calls setFeatureState with the right id and value', () => {
-    const { map, fire } = makeMapWithEvents();
-    jest.mocked(maplibregl.Map).mockImplementationOnce(() => {
-      return map as never;
-    });
-    const adapter = createMapLibreAdapter();
-    adapter.mount(
-      makeContainer(),
-      baseSpec([
-        {
-          mapDataId: 'pop',
-          mapId: 'states',
-          data: [{ geometryId: 'BR', value: 211 }],
-        },
-      ]),
-      'v'
-    );
-    fire('load');
-    jest.mocked(map.setFeatureState).mockClear();
-
-    adapter.applyPatch?.({
-      target: 'mapData',
-      op: 'replace',
-      path: 'mapData.pop.data.BR',
-      value: 215,
-    });
-
-    expect(map.setFeatureState).toHaveBeenCalledWith(
-      { source: 'states', id: 'BR' },
-      { value: 215 }
-    );
-  });
-
-  // 3.6
-  test('mounting without mapData never calls setFeatureState (V1 no-regression)', () => {
-    const { map, fire } = makeMapWithEvents();
-    jest.mocked(maplibregl.Map).mockImplementationOnce(() => {
-      return map as never;
-    });
-    const adapter = createMapLibreAdapter();
-    adapter.mount(makeContainer(), baseSpec(), 'v');
-
-    fire('load');
-
-    expect(map.setFeatureState).not.toHaveBeenCalled();
-  });
-
-  // 3.7 — joinKey: applies setFeatureState directly by geometryId (promoted to feature.id)
-  test('joinKey: applies setFeatureState directly by geometryId', () => {
-    const { map, fire } = makeMapWithEvents();
-    jest.mocked(maplibregl.Map).mockImplementationOnce(() => {
-      return map as never;
-    });
-    const adapter = createMapLibreAdapter();
-    adapter.mount(
-      makeContainer(),
-      baseSpec([
-        {
-          mapDataId: 'pop',
-          mapId: 'states',
-          joinKey: 'name',
-          data: [
-            { geometryId: 'BR', value: 211 },
-            { geometryId: 'AR', value: 45 },
-          ],
-        },
-      ]),
-      'v'
-    );
-
-    fire('load');
-
-    // `joinKey` is promoted to `feature.id`, so the join value is the id —
-    // no viewport-bound resolution.
-    expect(map.querySourceFeatures).not.toHaveBeenCalled();
-    expect(map.setFeatureState).toHaveBeenCalledWith(
-      { source: 'states', id: 'BR' },
-      { value: 211 }
-    );
-    expect(map.setFeatureState).toHaveBeenCalledWith(
-      { source: 'states', id: 'AR' },
-      { value: 45 }
-    );
-  });
-
-  // 3.7b — joinKey: numeric-string geometryId is coerced to a number id
-  test('joinKey: coerces numeric-string geometryId to a number id', () => {
-    const { map, fire } = makeMapWithEvents();
-    jest.mocked(maplibregl.Map).mockImplementationOnce(() => {
-      return map as never;
-    });
-
-    const adapter = createMapLibreAdapter();
-    adapter.mount(
-      makeContainer(),
-      baseSpec([
-        {
-          mapDataId: 'pop',
-          mapId: 'states',
-          joinKey: 'district_code',
-          data: [
-            { geometryId: '67', value: 211 },
-            { geometryId: '88', value: 45 },
-          ],
-        },
-      ]),
-      'v'
-    );
-
-    fire('load');
-
-    expect(map.setFeatureState).toHaveBeenCalledWith(
-      { source: 'states', id: 67 },
-      { value: 211 }
-    );
-    expect(map.setFeatureState).toHaveBeenCalledWith(
-      { source: 'states', id: 88 },
-      { value: 45 }
-    );
-  });
-
-  // 3.4 — NaN/Infinity row values are sanitized to 0 before setFeatureState
-  test.each([
-    ['NaN', Number.NaN],
-    ['Infinity', Number.POSITIVE_INFINITY],
-    ['-Infinity', Number.NEGATIVE_INFINITY],
-  ])(
-    'non-finite value (%s) is sanitized to 0 before setFeatureState (no joinKey)',
-    (_label, badValue) => {
-      const { map, fire } = makeMapWithEvents();
-      jest.mocked(maplibregl.Map).mockImplementationOnce(() => {
-        return map as never;
-      });
-      const adapter = createMapLibreAdapter();
-      adapter.mount(
-        makeContainer(),
-        baseSpec([
-          {
-            mapDataId: 'pop',
-            mapId: 'states',
-            data: [{ geometryId: 'BR', value: badValue }],
-          },
-        ]),
-        'v'
-      );
-
-      fire('load');
-
-      expect(map.setFeatureState).toHaveBeenCalledWith(
-        { source: 'states', id: 'BR' },
-        { value: 0 }
-      );
-    }
-  );
-
-  // 3.4b — granular replace also sanitizes NaN/Infinity
-  test('granular replace patch sanitizes non-finite value to 0', () => {
-    const { map, fire } = makeMapWithEvents();
-    jest.mocked(maplibregl.Map).mockImplementationOnce(() => {
-      return map as never;
-    });
-    const adapter = createMapLibreAdapter();
-    adapter.mount(
-      makeContainer(),
-      baseSpec([
-        {
-          mapDataId: 'pop',
-          mapId: 'states',
-          data: [{ geometryId: 'BR', value: 100 }],
-        },
-      ]),
-      'v'
-    );
-    fire('load');
-    jest.mocked(map.setFeatureState).mockClear();
-
-    adapter.applyPatch?.({
-      target: 'mapData',
-      op: 'replace',
-      path: 'mapData.pop.data.BR',
-      value: Number.NaN,
-    });
-
-    expect(map.setFeatureState).toHaveBeenCalledWith(
-      { source: 'states', id: 'BR' },
-      { value: 0 }
-    );
-  });
-
-  // 3.8 — joinKey: granular replace applies directly by geometryId
-  test('joinKey: granular replace patch applies directly by geometryId', () => {
-    const { map, fire } = makeMapWithEvents();
-    jest.mocked(maplibregl.Map).mockImplementationOnce(() => {
-      return map as never;
-    });
-    const adapter = createMapLibreAdapter();
-    adapter.mount(
-      makeContainer(),
-      baseSpec([
-        {
-          mapDataId: 'pop',
-          mapId: 'states',
-          joinKey: 'name',
-          data: [{ geometryId: 'BR', value: 211 }],
-        },
-      ]),
-      'v'
-    );
-    fire('load');
-    jest.mocked(map.setFeatureState).mockClear();
-
-    adapter.applyPatch?.({
-      target: 'mapData',
-      op: 'replace',
-      path: 'mapData.pop.data.BR',
-      value: 999,
-    });
-
-    expect(map.querySourceFeatures).not.toHaveBeenCalled();
-    expect(map.setFeatureState).toHaveBeenCalledWith(
-      { source: 'states', id: 'BR' },
-      { value: 999 }
-    );
-  });
-
-  // 3.9 — joinKey: granular replace no longer depends on the source being loaded
-  test('joinKey: granular replace applies even when source is not loaded', () => {
-    const { map, fire } = makeMapWithEvents();
-    jest.mocked(map.isSourceLoaded).mockReturnValue(false);
-    jest.mocked(maplibregl.Map).mockImplementationOnce(() => {
-      return map as never;
-    });
-    const adapter = createMapLibreAdapter();
-    adapter.mount(
-      makeContainer(),
-      baseSpec([
-        {
-          mapDataId: 'pop',
-          mapId: 'states',
-          joinKey: 'name',
-          data: [{ geometryId: 'BR', value: 211 }],
-        },
-      ]),
-      'v'
-    );
-    fire('load');
-    jest.mocked(map.setFeatureState).mockClear();
-
-    adapter.applyPatch?.({
-      target: 'mapData',
-      op: 'replace',
-      path: 'mapData.pop.data.BR',
-      value: 999,
-    });
-
-    // Feature state is keyed by id (no querySourceFeatures), so it applies
-    // lazily regardless of load state — MapLibre paints it when the tile loads.
-    expect(map.setFeatureState).toHaveBeenCalledWith(
-      { source: 'states', id: 'BR' },
-      { value: 999 }
-    );
-  });
-
-  // 3.10 — pending listener is cancelled on op:remove before source loads
-  test('op:remove cancels pending sourcedata listener; no setFeatureState when source eventually loads', () => {
-    const { map, fire } = makeMapWithEvents();
-    jest.mocked(map.isSourceLoaded).mockReturnValue(false);
-    jest.mocked(maplibregl.Map).mockImplementationOnce(() => {
-      return map as never;
-    });
-    const adapter = createMapLibreAdapter();
-    adapter.mount(
-      makeContainer(),
-      baseSpec([
-        {
-          mapDataId: 'pop',
-          mapId: 'states',
-          data: [{ geometryId: 'BR', value: 1 }],
-        },
-      ]),
-      'v'
-    );
-    fire('load'); // registers pending sourcedata listener
-
-    adapter.applyPatch?.({ target: 'mapData', op: 'remove', value: 'pop' });
-
-    // Source loads after removal — stale listener must NOT apply feature state
-    fire('sourcedata', { sourceId: 'states', isSourceLoaded: true });
-
-    expect(map.setFeatureState).not.toHaveBeenCalled();
-  });
-
-  // 3.11 — pending listener is replaced on full-entry replace before source loads
-  test('full-entry replace cancels stale listener and applies only new data when source loads', () => {
-    const { map, fire } = makeMapWithEvents();
-    jest.mocked(map.isSourceLoaded).mockReturnValue(false);
-    jest.mocked(maplibregl.Map).mockImplementationOnce(() => {
-      return map as never;
-    });
-    const adapter = createMapLibreAdapter();
-    adapter.mount(
-      makeContainer(),
-      baseSpec([
-        {
-          mapDataId: 'pop',
-          mapId: 'states',
-          data: [{ geometryId: 'BR', value: 1 }],
-        },
-      ]),
-      'v'
-    );
-    fire('load'); // registers pending sourcedata listener for old data (value: 1)
-
-    adapter.applyPatch?.({
-      target: 'mapData',
-      op: 'replace',
-      path: 'mapData.pop',
-      value: {
-        mapDataId: 'pop',
-        mapId: 'states',
-        data: [{ geometryId: 'BR', value: 99 }],
-      },
-    });
-
-    // Source loads — only the replacement data (99) must be applied, not the stale (1)
-    fire('sourcedata', { sourceId: 'states', isSourceLoaded: true });
-
-    expect(map.setFeatureState).toHaveBeenCalledTimes(1);
-    expect(map.setFeatureState).toHaveBeenCalledWith(
-      { source: 'states', id: 'BR' },
-      { value: 99 }
-    );
-  });
-
-  test('update() with changed mapData reapplies active legend fill-color expression', () => {
-    const { map, fire } = makeMapWithEvents();
-    jest.mocked(map.isStyleLoaded).mockReturnValue(true);
-    jest.mocked(map.getLayer).mockReturnValue({} as never);
-    jest.mocked(maplibregl.Map).mockImplementationOnce(() => {
-      return map as never;
-    });
-
-    const adapter = createMapLibreAdapter();
-    const withLegend = baseSpec([
-      {
-        mapDataId: 'pop',
-        mapId: 'states',
-        data: [{ geometryId: 'BR', value: 211 }],
-      },
-    ]);
-    withLegend.layers = [
-      {
-        ...withLegend.layers[0],
-        legends: [
-          {
-            id: 'status',
-            colorBy: {
-              type: 'categorical',
-              property: 'status',
-              mapping: { high: '#16a34a' },
-              defaultColor: '#6b7280',
-            },
-          },
-        ],
-        activeLegendId: 'status',
-      },
-    ];
-
-    adapter.mount(makeContainer(), withLegend, 'v');
-    fire('load');
-    jest.mocked(map.setPaintProperty).mockClear();
-
-    adapter.update({
-      ...withLegend,
-      mapData: [
-        {
-          mapDataId: 'pop',
-          mapId: 'states',
-          data: [{ geometryId: 'BR', value: 300 }],
-        },
-      ],
-    });
-
-    expect(map.setPaintProperty).toHaveBeenCalledWith('fill', 'fill-color', [
-      'match',
-      ['to-string', ['coalesce', ['feature-state', 'value'], '__missing__']],
-      'high',
-      '#16a34a',
-      '#6b7280',
-    ]);
-  });
-
-  test('update() keeps adapter-managed legend fill-color for mapData layers', () => {
-    const { map, fire } = makeMapWithEvents();
-    jest.mocked(map.isStyleLoaded).mockReturnValue(true);
-    jest.mocked(map.getLayer).mockReturnValue({} as never);
-    jest.mocked(maplibregl.Map).mockImplementationOnce(() => {
-      return map as never;
-    });
-
-    const adapter = createMapLibreAdapter();
-    const withLegend = baseSpec([
-      {
-        mapDataId: 'pop',
-        mapId: 'states',
-        data: [{ geometryId: 'BR', value: 211 }],
-      },
-    ]);
-    withLegend.layers = [
-      {
-        ...withLegend.layers[0],
-        mapDataId: 'pop',
-        activeLegendId: 'population',
-      },
-    ];
-    withLegend.legends = [
-      {
-        id: 'population',
-        colorBy: {
-          type: 'quantitative',
-          property: 'population',
-          scale: 'threshold',
-          thresholds: [50, 100],
-          colors: ['#f0f9ff', '#bfdbfe', '#60a5fa'],
-          defaultColor: '#f0f9ff',
-        },
-      },
-    ];
-
-    adapter.mount(makeContainer(), withLegend, 'v');
-    fire('load');
-    jest.mocked(map.setPaintProperty).mockClear();
-
-    adapter.update({
-      ...withLegend,
-      legends: [
-        {
-          id: 'population',
-          colorBy: {
-            type: 'quantitative',
-            property: 'population',
-            scale: 'threshold',
-            thresholds: [50, 100],
-            colors: ['#eff6ff', '#93c5fd', '#3b82f6'],
-            defaultColor: '#eff6ff',
-          },
-        },
-      ],
-    });
-
-    expect(map.setPaintProperty).toHaveBeenCalledWith('fill', 'fill-color', [
-      'step',
-      ['to-number', ['coalesce', ['feature-state', 'value'], 0], 0],
-      '#eff6ff',
-      50,
-      '#93c5fd',
-      100,
-      '#3b82f6',
-    ]);
-  });
-
-  test('mapData patch replace reapplies active legend fill-color expression', () => {
-    const { map, fire } = makeMapWithEvents();
-    jest.mocked(map.isStyleLoaded).mockReturnValue(true);
-    jest.mocked(map.getLayer).mockReturnValue({} as never);
-    jest.mocked(maplibregl.Map).mockImplementationOnce(() => {
-      return map as never;
-    });
-
-    const adapter = createMapLibreAdapter();
-    const withLegend = baseSpec([
-      {
-        mapDataId: 'pop',
-        mapId: 'states',
-        data: [{ geometryId: 'BR', value: 211 }],
-      },
-    ]);
-    withLegend.layers = [
-      {
-        ...withLegend.layers[0],
-        legends: [
-          {
-            id: 'status',
-            colorBy: {
-              type: 'categorical',
-              property: 'status',
-              mapping: { high: '#16a34a' },
-              defaultColor: '#6b7280',
-            },
-          },
-        ],
-        activeLegendId: 'status',
-      },
-    ];
-
-    adapter.mount(makeContainer(), withLegend, 'v');
-    fire('load');
-    jest.mocked(map.setPaintProperty).mockClear();
-
-    adapter.applyPatch?.({
-      target: 'mapData',
-      op: 'replace',
-      path: 'mapData.pop.data.BR',
-      value: 500,
-    });
-
-    expect(map.setPaintProperty).toHaveBeenCalledWith('fill', 'fill-color', [
-      'match',
-      ['to-string', ['coalesce', ['feature-state', 'value'], '__missing__']],
-      'high',
-      '#16a34a',
-      '#6b7280',
-    ]);
-  });
-
-  test('mapData patch replace reapplies legend fill-color when layer appears after initial pass', () => {
-    const { map, fire } = makeMapWithEvents();
-    jest.mocked(map.isStyleLoaded).mockReturnValue(true);
-    jest.mocked(map.getLayer).mockReturnValue(null);
-    jest.mocked(maplibregl.Map).mockImplementationOnce(() => {
-      return map as never;
-    });
-
-    const adapter = createMapLibreAdapter();
-    const withLegend = baseSpec([
-      {
-        mapDataId: 'pop',
-        mapId: 'states',
-        data: [{ geometryId: 'BR', value: 211 }],
-      },
-    ]);
-    withLegend.layers = [
-      {
-        ...withLegend.layers[0],
-        legends: [
-          {
-            id: 'status',
-            colorBy: {
-              type: 'categorical',
-              property: 'status',
-              mapping: { high: '#16a34a' },
-              defaultColor: '#6b7280',
-            },
-          },
-        ],
-        activeLegendId: 'status',
-      },
-    ];
-
-    adapter.mount(makeContainer(), withLegend, 'v');
-    fire('load');
-    jest.mocked(map.setPaintProperty).mockClear();
-
-    adapter.applyPatch?.({
-      target: 'mapData',
-      op: 'replace',
-      path: 'mapData.pop.data.BR',
-      value: 500,
-    });
-
-    expect(map.setPaintProperty).not.toHaveBeenCalled();
-
-    jest.mocked(map.getLayer).mockReturnValue({} as never);
-    fire('styledata', { dataType: 'style' });
-
-    expect(map.setPaintProperty).toHaveBeenCalledWith('fill', 'fill-color', [
-      'match',
-      ['to-string', ['coalesce', ['feature-state', 'value'], '__missing__']],
-      'high',
-      '#16a34a',
-      '#6b7280',
-    ]);
   });
 });
